% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin}
\alias{admin}
\title{Admin Console}
\usage{
admin()
}
\description{
Launches the administration console.
}
\details{
The Admin Console is a standalone Shiny web app delivered along with
the package to administrate the items of the project.

It will check the \code{R_KITEMS_PATH} environment variable and look
for the YAML config file in the procided path.
}
\examples{
\dontrun{
# set environment (where to find the _kitems.yml)
Sys.setenv("R_KITEMS_PATH" = "D:/data")

# launch the Admin Console
admin()
}
}
