% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_item_layouts.R
\name{admin_item_layout}
\alias{admin_item_layout}
\title{Item Layout}
\usage{
admin_item_layout(x)
}
\arguments{
\item{x}{the item config}
}
\value{
a bslib::nav_panel() object
}
\description{
This is a UI / Layout function for the Admin Console
}
\examples{
\dontrun{
admin_item_layout(item_config)
}
}
\keyword{internal}
