% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_layouts.R
\name{admin_layout}
\alias{admin_layout}
\title{Admin Console Layout}
\usage{
admin_layout()
}
\value{
an HTML tag.
}
\description{
This is a UI / Layout function for the Admin Console.
}
\details{
The main layout will be used to dynamically insert / remove content
from the server side.
}
\examples{
\dontrun{
admin_layout()
}
}
\keyword{internal}
