% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_migration_layouts.R
\name{admin_migration_required_layout}
\alias{admin_migration_required_layout}
\title{Migration Required Layout}
\usage{
admin_migration_required_layout(files)
}
\arguments{
\item{files}{a list of data.model files to migrate}
}
\value{
a layout
}
\description{
Migration Required Layout
}
\examples{
\dontrun{
admin_migration_required_layout("./item_data_model.rds")
}
}
\keyword{internal}
