% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_console.R
\name{admin_server}
\alias{admin_server}
\title{Admin Console Server}
\usage{
admin_server(input, output, session)
}
\arguments{
\item{input}{the shiny input object}

\item{output}{the shiny output object}

\item{session}{the shiny session object}
}
\value{
nothing.
}
\description{
Admin Console Server
}
\examples{
\dontrun{
admin_server(input, output, session)
}
}
\keyword{internal}
