% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_widget.R
\name{admin_widget}
\alias{admin_widget}
\title{Admin Widget}
\usage{
admin_widget(id)
}
\arguments{
\item{id}{the id used to launch the admin module server}
}
\value{
A \code{htmltools::tagList()} object containing the admin UI
}
\description{
Admin Widget
}
\examples{
\dontrun{
admin_widget(id = "mydata")
}
}
