% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_layouts.R
\name{admin_yaml_message}
\alias{admin_yaml_message}
\title{YAML Messages}
\usage{
admin_yaml_message(config)
}
\arguments{
\item{config}{a YAML config}
}
\value{
an HTML tag
}
\description{
This is a helper function to return messages about the YAML config.
}
\examples{
\dontrun{
admin_yaml_message(config)
}
}
\keyword{internal}
