% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask Attribute(s)}
\usage{
ask(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to include / reset (quoted or unquoted)}
}
\value{
a data.frame of the updated data model or a character vector
}
\description{
Set attribute(s) to include in the item form.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the attribute name(s).
}
\examples{
ask(data_model(c(foo = "character"), skip = TRUE), foo)
ask(data_model(c(foo = "character"), skip = TRUE, refresh = TRUE), "foo")
ask(data_model(c(foo = "character", bar = "logical"), skip = TRUE), "foo", bar)

# getter
ask(data_model(c(foo = "character")))
}
\seealso{
\code{\link[=avoid]{avoid()}} \code{\link[=attribute_update]{attribute_update()}}
}
