% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_create.R
\name{attribute_create}
\alias{attribute_create}
\title{Add Attribute}
\usage{
attribute_create(data.model = NULL, name, class, ...)
}
\arguments{
\item{data.model}{a data model, structured as an output of the \code{data_model()} function}

\item{name}{a character vector for the attribute name to add}

\item{class}{a character vector for the attribute class to add}

\item{...}{further arguments passed to \code{data_model()}}
}
\value{
The updated data model (or \code{data.model} if it fails)
}
\description{
Add an attribute to a data model
}
\details{
Multiple attribute creation is supported.
}
\examples{
\dontrun{

# -- Add single attribute
attribute_create(data.model = NULL, name = "new_attribute", class = "character")
attribute_create(data.model = NULL, name = "total", class = "numeric", default.val = 0)
attribute_create(data.model = NULL, name = "date", class = "Date", default.fun = "Sys.Date")
attribute_create(data.model = NULL, name = "progress", class = "integer", skip = "progress")
attribute_create(data.model = NULL, name = "internal", class = "logical", display = "internal")

# -- create multiple attributes
attribute_create(data.model = NULL, name = c("foo", "bar"), class = c("character", "numeric"))

}
}
\seealso{
\code{\link[=data_model]{data_model()}}
}
