% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_delete.R
\name{attribute_delete}
\alias{attribute_delete}
\title{Delete Attribute}
\usage{
attribute_delete(data.model, name, items)
}
\arguments{
\item{data.model}{the data.model to update}

\item{name}{the name of the attribute to delete}

\item{items}{the items to update}
}
\value{
a named list(data.model, items)
}
\description{
Delete an attribute from the data.model & items
}
\details{
The function returns a list so that both data.model & items
are treated in a single place. This ensures data integrity.
}
\examples{
\dontrun{
attribute_delete(dm, items, name = "comment")
}

}
