% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_input.R
\name{attribute_input}
\alias{attribute_input}
\title{Attribute Input}
\usage{
attribute_input(
  name,
  type,
  value = NULL,
  choices = NULL,
  create = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{name}{the name of the attribute.}

\item{type}{the type of the attribute.}

\item{value}{the value to be used to initialize the input.}

\item{choices}{a list of values to select from (see details).}

\item{create}{a logical (default = \code{FALSE}) if user is allowed to create values (see details).}

\item{session}{optional, the shiny session object.}
}
\value{
An input that can be added to the UI definition.
}
\description{
Build input for an attribute.
}
\details{
By default (\code{choices = NULL}), the function will return an input driven by the
type of the attribute.
When \code{choices} are provided, it will be replaced by a selectizeInput. User
will be allowed to create additional values depending on \code{create} (otherwise ignored).
}
\examples{
\dontrun{
# -- create inputs
attribute_input(name = "total", type = "numeric", value = 10)
}
}
