% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_input_update.R
\name{attribute_input_update}
\alias{attribute_input_update}
\title{Update Attribute Input}
\usage{
attribute_input_update(data.model, shortcut_trigger, MODULE = NULL)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{shortcut_trigger}{basically the value of input$shortcut_trigger}

\item{MODULE}{an optional character string for the trace}
}
\description{
Update Attribute Input
}
\examples{
\dontrun{
attribute_input_update(
data.model = mydata$data_model,
shortcut_trigger = "name_banana",
MODULE = "(mydata)")
}
}
\keyword{internal}
