% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_shortcut.R
\name{attribute_shortcut}
\alias{attribute_shortcut}
\title{Attribute Shortcuts}
\usage{
attribute_shortcut(colClass, suggestions, ns)
}
\arguments{
\item{colClass}{a length-one named vector. \code{names(colClass)} is the name of the attribute,
and \code{colClass} is the type (class) of the attribute.}

\item{suggestions}{a list of suggestions, output of \code{\link[=attribute_suggestion]{attribute_suggestion()}}}

\item{ns}{the namespace function reference}
}
\value{
a list of actionLink objects

The return value is the output of \link[base]{lapply} applying \link[shiny]{actionLink} over \code{suggestions}

The actionLink has an \code{onclick} property that will trigger \code{input$shortcut_trigger} (\code{ns(shortcut_trigger)})
Its value is of the form \code{[ns]-[attribute]_[value]}
Basically, applying \code{tail(unlist(strsplit(input$shortcut_trigger, split = "-")), 1)} will access attribute_values.

Note that for POSIXct attribute, the shortcut_trigger input will not carry the timezone information.
Clicking on the corresponding actionLink will only trigger date & time inputs update.
}
\description{
Build attribute shortcuts list
}
\examples{
\dontrun{
attribute_shortcut(colClass = c(name = "character"),
suggestions = list(mango = 25, banana = 12, lemon = 10),
ns = session$ns)
}
}
\seealso{
\code{\link[=attribute_suggestion]{attribute_suggestion()}} \code{\link[=item_form]{item_form()}}
}
