% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_value.R
\name{attribute_value}
\alias{attribute_value}
\title{Turn Value(s) into Attribute}
\usage{
attribute_value(key, value, data.model)
}
\arguments{
\item{key}{the name of the attribute}

\item{value}{a vector of input values}

\item{data.model}{the data.frame of the data.model}
}
\value{
a vector, with values checked against the data.model
}
\description{
Turn Value(s) into Attribute
}
\examples{
\dontrun{
attribute_value(key, value, data.model)}
}
