% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoid.R
\name{avoid}
\alias{avoid}
\title{Avoid Attribute}
\usage{
avoid(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to avoid(quoted or unquoted)}
}
\value{
a data.frame of the updated data model or a character vector
}
\description{
Defines attribute(s) to avoid in the item form.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the attribute name(s).
}
\examples{
avoid(data_model(c(foo = "character", bar = "logical")), bar)
avoid(data_model(c(foo = "character", bar = "logical")), foo, "bar")

# getter
avoid(data_model(c(foo = "character", bar = "logical")))
}
\seealso{
\code{\link[=refresh]{refresh()}} \code{\link[=attribute_update]{attribute_update()}}
}
