% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{Backup Files}
\usage{
backup(
  type = c("config", "items", "dm"),
  id = NULL,
  max = 1,
  path = Sys.getenv("R_KITEMS_PATH")
)
}
\arguments{
\item{type}{the type of file to backup (see details).}

\item{id}{the id of the item or data model.}

\item{max}{an integer (default = 1) to indicate how many backup files are allowed.}

\item{path}{optional, the path to the data.}
}
\description{
Backup config, data model or items files.
}
\details{
\code{type} accepts the following values:
\itemize{
\item "config" (the default)
\item "items"
\item "dm"
}

"dm" is kept to enable data.model backup before migration to the YAML config.

\code{id} will be ignored if \code{type = "config"}

The recommended way to define the \code{path} argument is to set the \code{R_KITEMS_PATH}
environment variable.

Backup file will be named as \emph{_kitems_YYYY-MM-DD.yml},
\emph{id_data_model_YYYY-MM-DD.rds} or \emph{id_items_YYYY-MM-DD.csv}
If same file already exists, it will be overwritten.
}
\examples{
\dontrun{
# backup config
backup()
}
}
