% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check Integrity}
\usage{
check(x, ...)
}
\arguments{
\item{x}{the object to check (data.model or items).}

\item{...}{parameters to pass to the dedicated check integrity function.}
}
\value{
the output of the dedicated check integrity function
}
\description{
Wrapper function to either check data.model or items integrity.
}
\details{
\code{x} will be send to the first argument of the dedicated check integrity function.
At least one argument is expected in ...
See dm_integrity() or item_integrity() functions to know what is expected
as ... argument(s)
}
\examples{
\dontrun{
data_model |> check(items = foo, fix = FALSE)
items |> check(data.model = foo, fix = TRUE)
}
}
