% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{clicked_column_name}
\alias{clicked_column_name}
\title{Clicked column name}
\usage{
clicked_column_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Clicked column name
}
\examples{
clicked_column_name(id = "mydata")
}
