% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_append}
\alias{config_attribute_append}
\title{Append Attribute}
\usage{
config_attribute_append(
  config,
  item,
  attribute,
  hide = FALSE,
  skip = FALSE,
  refresh = FALSE
)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name of the target item}

\item{attribute}{the attribute config}

\item{hide}{a logical (default = FALSE) if the attribute should be hidden}

\item{skip}{a logical (default = FALSE) if the attribute should be skipped}

\item{refresh}{a logical (default = FALSE) if the attribute should be refreshed}
}
\value{
the new config
}
\description{
Append Attribute
}
\examples{
\dontrun{
config_attribute_append(config, item = foo,
                        attribute = config_attribute_create(name = "bar", type "numeric"))
}
}
