% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_behavior}
\alias{config_attribute_behavior}
\title{Attribute Behaviors}
\usage{
config_attribute_behavior(config, item, behavior, ..., set = TRUE)
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}

\item{behavior}{which behavior to manipulate ("skip", "refresh" or "hide")}

\item{...}{the name of the attribute(s) to skip}

\item{set}{a logical (default = TRUE) if the behavior should be set or unset}
}
\value{
a character vector
}
\description{
Function to set/unset the skip, refresh & hide attribute behaviors.
}
\examples{
\dontrun{
config_attribute_behavior(config, item = "foo", behavior = "skip", "id")
config_attribute_behavior(config, item = "foo", behavior = "hide", "id", set = FALSE)
}
}
