% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_create}
\alias{config_attribute_create}
\title{Create Attribute}
\usage{
config_attribute_create(name, type)
}
\arguments{
\item{name}{the name of the attribute}

\item{type}{the type of the attribute}
}
\value{
a list
}
\description{
Create Attribute
}
\examples{
config_attribute_create(name = "att_1", type = "integer")
}
