% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_drop}
\alias{config_attribute_drop}
\title{Drop Attribute}
\usage{
config_attribute_drop(config, item, attribute)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the target item}

\item{attribute}{the name of the attribute to drop}
}
\value{
the new yaml config
}
\description{
Drop Attribute
}
\examples{
\dontrun{
config_attribute_drop(config, item = "foo", attribute = "att_1")
}
}
