% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_move}
\alias{config_attribute_move}
\title{Move Attribute}
\usage{
config_attribute_move(config, item, attribute, where)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}

\item{attribute}{the name of the attribute to move}

\item{where}{a list(position = c("before", "after"), attribute = "foo")}
}
\value{
the new config
}
\description{
Move Attribute
}
\examples{
\dontrun{
config_attribute_move(config, "item_1", "att_1", list(position = "after", attribute = "att_2"))
}
}
