% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_position}
\alias{config_attribute_position}
\title{Attribute Position}
\usage{
config_attribute_position(config, item, attribute)
}
\arguments{
\item{config}{a yaml config}

\item{item}{the name (id) of the item group}

\item{attribute}{the name of the attribute to locate}
}
\value{
an integer
}
\description{
Attribute Position
}
\examples{
\dontrun{
config_attribute_position(config, item = "foo", attribute = "bar")
}
}
