% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_update}
\alias{config_attribute_update}
\title{Update Attribute}
\usage{
config_attribute_update(config, item, attribute)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}

\item{attribute}{the attribute config}
}
\value{
the new config
}
\description{
Update Attribute
}
\examples{
\dontrun{
config_attribute_update(config, item = foo,
                        attribute = config_attribute_create(name = "bar", type "numeric"))
}
}
