% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_check.R
\name{config_check}
\alias{config_check}
\title{Check Config}
\usage{
config_check(config, path = Sys.getenv("R_KITEMS_PATH"))
}
\arguments{
\item{config}{a config list}

\item{path}{an option path}
}
\value{
a list
}
\description{
Check YAML config structure & content
}
\details{
The function will return a list of problems or an empty list.
Each problem has a type (note, warning or error) and a message.
Errors are most likely to prevent the package from working properly.
Warnings may affect the behavior.
Notes are here for information but should not affect the behavior.

By default, path is initialized with the \code{R_KITEMS_PATH} environment
variable.
}
\examples{
\dontrun{
config_check(config)
}
}
\keyword{internal}
