% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_create}
\alias{config_create}
\title{Create YAML Configuration File}
\usage{
config_create(project)
}
\arguments{
\item{project}{a character string to indicate the name of the project}
}
\value{
a list
}
\description{
Create YAML Configuration File
}
\examples{
config_create(project = "my_project")
}
