% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_extract}
\alias{config_extract}
\title{Extract Config}
\usage{
config_extract(config, item = NULL, attribute = NULL)
}
\arguments{
\item{config}{the yaml configuration}

\item{item}{optional, the name (id) of the item}

\item{attribute}{optional, the name of the attribute}
}
\value{
a list (the yaml section for the attribute)
}
\description{
Extract the section for a specific item or attribute.
}
\details{
When \code{item} is NULL, it is expected that \code{config} contains the yaml
section for an item.

If both \code{item} and \code{attribute} are left NULL, the function will
return the input.
}
\examples{
\dontrun{
config_extract(config, item = "foo")
config_extract(config, item = "foo", attribute = "bar")
}
}
