% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_append}
\alias{config_item_append}
\title{Append Item}
\usage{
config_item_append(config, ...)
}
\arguments{
\item{config}{the yaml config}

\item{...}{item(s) to append}
}
\value{
the new config
}
\description{
Append Item
}
\examples{
config_item_append(config_create("prj"),
                   config_item_create(id = "a", path = "."))
}
