% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_behavior}
\alias{config_item_behavior}
\title{Item Behaviors}
\usage{
config_item_behavior(config, item, behavior = c("skip", "refresh", "hide"))
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}

\item{behavior}{the name of the behavior (default = "skip")}
}
\value{
a character vector or NULL
}
\description{
Get the attribute names for a specific behavior.
}
\details{
Behaviors:
\itemize{
\item "skip" the default
\item "refresh"
\item "hide"
}
}
\examples{
\dontrun{
config_item_behavior(config, item = "foo")
}
}
