% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_colclasses}
\alias{config_item_colclasses}
\title{Item colClasses}
\usage{
config_item_colclasses(config, item)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}
}
\value{
a named vector or list() if the item is not found.
}
\description{
Item colClasses
}
\examples{
# create a config
yaml <- config_create("test") |>
  config_item_append(config_item_create(id = "foo", path = "/temp")) |>
  config_attribute_append(item = "foo",
                          attribute = config_attribute_create(name = "id",
                                                              type = "numeric"))

# get item colClasses
config_item_colclasses(yaml, "foo")
}
