% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_connector}
\alias{config_item_connector}
\title{Item Connector}
\usage{
config_item_connector(config, item)
}
\arguments{
\item{config}{the YAML config}

\item{item}{the name of the item}
}
\value{
a list
}
\description{
Item Connector
}
\examples{
yaml <- config_create("test") |>
  config_item_append(config_item_create(id = "foo", path = "/temp"))

config_item_connector(yaml, "foo")
}
