% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_create}
\alias{config_item_create}
\title{Create Item Config}
\usage{
config_item_create(id, description = NULL, path)
}
\arguments{
\item{id}{a character string to set the name of the item group}

\item{path}{a character string where to find the data}
}
\value{
a list
}
\description{
Helper function to create the config section for an item.
}
\examples{
config_item_create(id = "foo", path = "./")
}
