% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_drop}
\alias{config_item_drop}
\title{Drop Item}
\usage{
config_item_drop(config, item)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}
}
\value{
the new config
}
\description{
Drop Item
}
\examples{
\dontrun{
config_item_drop(config, item = "foo")
}
}
