% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_position}
\alias{config_item_position}
\title{Item Position}
\usage{
config_item_position(config, item)
}
\arguments{
\item{config}{a yaml config}

\item{item}{the name of the item to locate}
}
\value{
an integer
}
\description{
Item Position
}
\examples{
\dontrun{
config_item_position(config, item = "foo")
}
}
