% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_row_order}
\alias{config_item_row_order}
\title{Item Row Order}
\usage{
config_item_row_order(config, item)
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}
}
\value{
a character string or \code{NA} if \code{ìtem} is missing.
}
\description{
Getter function that returns the ordering info of an item.
}
\examples{
\dontrun{
config_item_row_order(config, item = "foo")
}
}
