% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_write.R
\name{config_write}
\alias{config_write}
\title{Write YAML config file}
\usage{
config_write(x, path = Sys.getenv("R_KITEMS_PATH"))
}
\arguments{
\item{x}{a list object to write to YAML}

\item{path}{where to store the _kitems YAML file}
}
\description{
Write YAML config file
}
\details{
By default, \code{path} uses the R_KITEMS_PATH environment variable.
}
\examples{
\dontrun{
config_write(list(foo = 1, bar = "two"))
}
}
\keyword{internal}
