% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Call Conversion Functions}
\usage{
convert(x, class, class.arg = NULL)
}
\arguments{
\item{x}{the object to be converted}

\item{class}{the target class}

\item{class.arg}{a character string with the arguments (see details)}
}
\value{
the converted object
}
\description{
Helper function to call the class conversion functions with arguments
}
\details{
\code{class.arg} must be like "list(arg1 = 12, arg2 = 'foo')"
}
\examples{
\dontrun{
convert("2025-09-10T13:16:55Z", "POSIXct", "list(format = '\%Y-\%m-\%dT\%H:\%M:\%S')")
}
}
\keyword{internal}
