% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons_ui.R
\name{create_BTN}
\alias{create_BTN}
\title{Create item button}
\usage{
create_BTN(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a rendered actionButton
}
\description{
Create item button
}
\examples{
create_BTN(id = "mydata")
}
