% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danger_zone_ui.R
\name{danger_zone_ui}
\alias{danger_zone_ui}
\title{Danger zone UI}
\usage{
danger_zone_ui(k_data_model, ns)
}
\arguments{
\item{k_data_model}{the reference of the data model reactive value}

\item{ns}{the namespace function}
}
\value{
a tagList object
}
\description{
Danger zone UI
}
\examples{
\dontrun{
danger_zone_ui(k_data_model, ns)
}
}
