% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danger_zone_widget.R
\name{danger_zone_widget}
\alias{danger_zone_widget}
\title{Danger Zone Widget}
\usage{
danger_zone_widget(k_data_model, ns)
}
\arguments{
\item{k_data_model}{the reference of the data model reactive value}

\item{ns}{the namespace function}
}
\value{
a tagList object
}
\description{
Danger Zone Widget
}
\examples{
\dontrun{
danger_zone_widget(k_data_model, ns)
}
}
