% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{data_model}
\alias{data_model}
\title{Data Model}
\usage{
data_model(
  colClasses,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  display = NULL,
  skip = NULL,
  refresh = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{colClasses}{a \emph{mandatory} named vector of classes, defining the data model.}

\item{default.val}{an optional named vector of values, defining the default values.}

\item{default.fun}{an optional named vector of functions, defining the default functions to be used to generate default values.}

\item{default.arg}{an optional named vector of arguments, to pass along with the default function.}

\item{display}{an optional character vector, indicating which attribute names should be displayed in the table view.}

\item{skip}{an optional character vector, indicating which attribute names should be skipped from the user input form.}

\item{refresh}{an optional named logical vector, to indicate if skipped attributes should be updated (see details).}

\item{sort.rank}{an optional named numeric vector, to define sort orders.}

\item{sort.desc}{an optional named logical vector, to define if sort should be descending.}
}
\value{
A data.frame containing the data model.
}
\description{
Build a data model
}
\details{
\code{colClasses} will be used to create the data.frame: names will define the attributes of the data model,
and values will define the class (type) of the attributes.

All \verb{default.*} parameters are optional. When provided, they will be used to match with names defined in colClasses:
\itemize{
\item order in those vectors doesn't matter
\item there is no need to set values for all attributes (see examples)
\item names in vectors not matching with colClasses names will be ignored
}

\code{default.fun} and \code{default.val} are mutual exclusive, with priority on \code{default.fun} (\code{default.val} will be forced to \code{NA})
\code{default.arg} requires \code{default.fun} not to be NULL (will be forced to \code{NA} otherwise)

\code{display} and \code{skip} directly contains the names of the attributes to set to TRUE

Attributes with \code{skip} set to TRUE will be ignored by default when an item is updated. To force the computation of a new
value (based on the default parameters), set \code{refresh} to TRUE as well.

If not provided, defaults will be applied:
\itemize{
\item \code{NA} for \code{default.val}, \code{default.fun} and \code{default.arg}
\item \code{FALSE} for \code{display}, \code{skip} and \code{refresh}
}
}
\examples{
# -- order in vectors doesn't matter:
default.val <- c("name" = "test", "total" = 2)
default.val <- c("total" = 2, "name" = "test")

# -- no need to set all values
colClasses <- c("id" = "numeric", "name" = "character", "total" = "numeric")
default.val <- c("name" = "test", "total" = 2)

# -- display and skip
display <- "id"
skip <- c("id", "date")

# -- sort
sort.rank = c("date" = 1, "total" = 2, "name" = 3)
sort.desc = c("date" = TRUE, "total" = FALSE)

data_model(colClasses, default.val, display = display, skip = skip)

}
