% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_ui.R
\name{date_slider_INPUT}
\alias{date_slider_INPUT}
\title{Date sliderInput}
\usage{
date_slider_INPUT(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a rendered sliderInput
}
\description{
Date sliderInput
}
\examples{
date_slider_INPUT(id = "mydata")
}
