% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_buttons_ui.R
\name{delete_BTN}
\alias{delete_BTN}
\title{Delete item button}
\usage{
delete_BTN(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a rendered actionButton
}
\description{
Delete item button
}
\examples{
delete_BTN(id = "mydata")
}
