% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{dialog}
\alias{dialog}
\title{Item Modal Dialog(s)}
\usage{
dialog(
  ...,
  workflow = c("create", "update", "delete"),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{the content to be displayed in the modal dialog.}

\item{workflow}{a character string to indicate workflow (see details).}

\item{session}{optional, the shiny session object.}
}
\value{
a modal dialog.
}
\description{
Produces a create / update / delete modal dialog.
}
\details{
Possible values for workflow are "create" (default), "update" or "delete".

\code{...} is typically the output of the \code{form()} function.
When \code{workflow = "delete"}, it will be ignored and replaced by a standard message.
}
\examples{
\dontrun{
dialog(workflow = "delete")
}
}
