% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display Attribute(s)}
\usage{
display(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to set as displayed (quoted or unquoted)}
}
\value{
a data.frame of the updated data model or a character vector
}
\description{
Basically the opposite of the \code{hide()} function
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the displayed attribute's names.
}
\examples{
display(data_model(c(foo = "character")), id)
display(data_model(c(foo = "character")), "id")
display(data_model(c(foo = "character", bar = "logical")), "foo", bar)

# getter
display(data_model(c(foo = "character")))
}
\seealso{
\code{\link[=hide]{hide()}} \code{\link[=attribute_update]{attribute_update()}}
}
