% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_apply_mask.R
\name{dm_apply_mask}
\alias{dm_apply_mask}
\title{Apply data model filter}
\usage{
dm_apply_mask(data.model, items)
}
\arguments{
\item{data.model}{a \emph{mandatory} data model}

\item{items}{a \emph{mandatory} item data.frame}
}
\value{
an item data.frame, without the filtered attributes (columns)
}
\description{
Apply data model filter
}
\examples{
\dontrun{
filetered_items <- dm_apply_mask(data.model = mydatamodel, items = myitems)
}
}
