% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_colClasses.R
\name{dm_colClasses}
\alias{dm_colClasses}
\title{Data Model colClasses}
\usage{
dm_colClasses(data.model)
}
\arguments{
\item{data.model}{a data.frame of the data model.}
}
\value{
A named vector of the attribute types.
}
\description{
Extract attribute classes from the data model.
}
\examples{
dm_colClasses(data.model = data_model(colClasses = c(name = "character", total = "numeric")))
}
