% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_default.R
\name{dm_default}
\alias{dm_default}
\title{Default Value(s)}
\usage{
dm_default(data.model)
}
\arguments{
\item{data.model}{the data.frame of the data model.}
}
\value{
A data.frame
}
\description{
Compute default value(s).
}
\details{
This function does not accept the data.model element of the config list as an input.
It should be turned into a data.frame first, using \code{\link[=yaml_to_dm]{yaml_to_dm()}}.

\code{data.model} should contain the following columns:
"name", "type", "default" & "values".
}
\examples{
dm <- data_model(colClasses = c(foo = "numeric"), default = 12)
dm_default(dm)
}
