% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_delete.R
\name{dm_delete}
\alias{dm_delete}
\title{Delete data model}
\usage{
dm_delete(data.model, items, dm_url, items_url, autosave, item.file)
}
\arguments{
\item{data.model}{the reference of the data model reactive value}

\item{items}{the reference of the items reactive value}

\item{dm_url}{the url (path + filename) of the data model}

\item{items_url}{the url (path + filename) of the items}

\item{autosave}{logical if the data should be saved}

\item{item.file}{a logical if the item file should be deleted}
}
\description{
Delete data model
}
\examples{
\dontrun{
dm_delete(data.model = mydata$data_model,
items = mydata$items,
dm_url,
items_url,
autosave = TRUE,
item.file = TRUE)
}
}
