% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_filter.R
\name{dm_display}
\alias{dm_display}
\title{Data Model Display}
\usage{
dm_display(data.model, set = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{set}{an optional character vector with the name of the attributes to set as hidden}
}
\value{
either the list of attributes that are hidden or an updated data model if set is not NULL
}
\description{
Data Model Display
}
\details{
If set is NULL, then the data model passed as data.model is returned
}
\examples{
\dontrun{
dm_display(data.model = mydatamodel, set = c("id", "internal"))
}
}
