% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_filter.R
\name{dm_filter}
\alias{dm_filter}
\title{Data model attribute filter}
\usage{
dm_filter(data.model, set = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{set}{an optional character vector with the name of the attributes to set as filtered}
}
\value{
either the list of attributes that are filtered or an updated data model if set is not NULL
}
\description{
Data model attribute filter
}
\details{
If filter is NULL, then the data model passed as data.model is returned
}
\examples{
\dontrun{
dm_filter(data.model = mydatamodel, filter = c("id", "internal"))
}
}
