% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_get_list.R
\name{dm_get_list}
\alias{dm_get_list}
\title{Available data models}
\usage{
dm_get_list(r)
}
\arguments{
\item{r}{the shared reactive communication object}
}
\value{
a list with the names (ids) of the data models available in session
}
\description{
Available data models
}
\examples{
\dontrun{
dm_get_list(r)
}
}
