% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_inputs_ui.R
\name{dm_inputs_ui}
\alias{dm_inputs_ui}
\title{Data model inputs UI}
\usage{
dm_inputs_ui(update = FALSE, attribute = NULL, names = NULL, types = NULL, ns)
}
\arguments{
\item{update}{a logical to turn ui into update mode}

\item{attribute}{a dataframe of the attribute to be updated}

\item{names}{a character vector of choices for attribute name}

\item{types}{a character vector of choices for attribute type}

\item{ns}{the namespace function}
}
\value{
a form, as a result from renderUI function
}
\description{
Data model inputs UI
}
\examples{
# -- create mode
dm_inputs_ui()

# -- update mode
\dontrun{
dm_inputs_ui(update = TRUE, attribute = dm[1, ])
}

}
