% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_mask.R
\name{dm_mask}
\alias{dm_mask}
\title{Data Model Mask}
\usage{
dm_mask(data.model)
}
\arguments{
\item{data.model}{a data.frame of the data model.}
}
\value{
A data.frame with renamed columns.
}
\description{
Apply mask on the data model.
}
\examples{
\dontrun{
dm_mask(data.model = dm)
}
}
