% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_migrate.R
\name{dm_migrate}
\alias{dm_migrate}
\title{Migrate data model}
\usage{
dm_migrate(data.model, name)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{name}{a named list of the missing columns (most probably a subset of the target colClasses)}
}
\value{
the migrated data.model
}
\description{
Migrate data model
}
\examples{
\dontrun{
dm_migrate(data.model, name = list(default.arg = "character"))
}
}
