% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{dm_name}
\alias{dm_name}
\title{Data Model Name}
\usage{
dm_name(id)
}
\arguments{
\item{id}{the id of the module server instance.}
}
\value{
The name of the corresponding data model.
}
\description{
Helper function to compute the data model name.
}
\examples{
dm_name(id = "mydata")
}
