% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_skip.R
\name{dm_skip}
\alias{dm_skip}
\title{Extract skipped attributes}
\usage{
dm_skip(data.model)
}
\arguments{
\item{data.model}{a data.frame containing the data model}
}
\value{
a vector containing the names of the attributes to skip (or NULL if data.model is NULL)
}
\description{
Extract skipped attributes
}
\examples{
\dontrun{
dm_skip(data.model = mydatamodel)
}
}
