% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_table_mask.R
\name{dm_table_mask}
\alias{dm_table_mask}
\title{Apply mask on data model table}
\usage{
dm_table_mask(data.model)
}
\arguments{
\item{data.model}{a data.frame of the data model}
}
\value{
a data.frame with renamed columns
}
\description{
Apply mask on data model table
}
\examples{
\dontrun{
dm_table_mask(data.model = dm)
}
}
