% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{dm_url}
\alias{dm_url}
\title{Data Model URL}
\usage{
dm_url(id, path = Sys.getenv("R_KITEMS_PATH"))
}
\arguments{
\item{id}{the id of the module server instance.}

\item{path}{the path to the data model and items.}
}
\value{
The URL of the data model
}
\description{
Data Model URL
}
\details{
The recommended way to define the \code{path} argument is to set the R_KITEMS_PATH
environment variable.
}
\examples{
dm_url(id = "mydata")
}
