% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_values.R
\name{dm_values}
\alias{dm_values}
\title{Attribute Values}
\usage{
dm_values(x, data = NULL)
}
\arguments{
\item{x}{a character string of the expression to evaluate}

\item{data}{a data.frame to use for tidy evaluation of \code{x}}
}
\value{
a vector of values
}
\description{
Compute attribute values with data-masking support
}
\examples{
dm_values(x = "suggest(1, 2)")
dm_values(x = "suggest(min(value))", data = data.frame(value = c(12, 10)))
}
