% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_sidebar.R
\name{dynamic_sidebar}
\alias{dynamic_sidebar}
\title{Generate dynamic menuItem}
\usage{
dynamic_sidebar(r)
}
\arguments{
\item{r}{the reactive shared communication object}
}
\value{
a sidebarMenu menuItem object with one menuSubItem per data model
}
\description{
Generate dynamic menuItem
}
\examples{
\dontrun{
dynamic_sidebar(r)
}
}
