% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{enforce}
\alias{enforce}
\title{Item Migration}
\usage{
enforce(items, name, type, fill = NA)
}
\arguments{
\item{items}{a data.frame of the items.}

\item{name}{a character string of the attribute name.}

\item{type}{a character string of the attribute type.}

\item{fill}{the value (default = \code{NA}) to be used to fill the existing rows (see details).}
}
\value{
The updated items data.frame.
}
\description{
Add attribute to existing items.
}
\details{
\code{fill} will be coerced to the class name provided in \code{type}.
If a vector is given as input for \code{fill}, it will be used to feed the created column.
Make sure the vector length is same as the number of rows, otherwise an error will be raised by R.
}
\examples{
\dontrun{
enforce(items = myitems, name = "comment", type = "character", fill = "none")
}
}
