% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Extend Item}
\usage{
extend(config, item = get_context(), ...)
}
\arguments{
\item{config}{the config object.}

\item{item}{the name (id) of the item group.}

\item{...}{one or several attribute instructions.}
}
\value{
a list.
}
\description{
Add specific attribute(s) with fine tuning.
}
\details{
This function is a wrapper around \code{design()} to provide a comprehensive
layered syntax.

It only understands the \code{attribute} instruction inside \code{...}:
\code{attribute = c(name = "note", type = "character")}
Anything else will be filtered.

When \code{item} is set in the parent frame, then the parameter can be skipped
in the function call.
}
\examples{
\dontrun{
# create config
config <- design(project = "test",
        item = "foo")

# set item (to skip it in following calls)
item <- "foo"

# extend
config |>
  extend(attribute = c(name = "date", type = "Date", default = "Sys.Date()"))

# is same as
config <- design(project = "test",
        item = "foo",
        attribute = c(name = "date", type = "Date", default = "Sys.Date()"))
}
}
\seealso{
\code{\link[=design]{design()}}, \code{\link[=parent.frame]{parent.frame()}}
}
