% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Input Values}
\usage{
extract(input, colClasses)
}
\arguments{
\item{input}{the input object from the shiny module.}

\item{colClasses}{a named vector of classes, defining the data model.}
}
\value{
a named list.
}
\description{
Extract the list of values from the input.
}
\details{
The output list will contain as many entries as the \code{colClasses} named vector.
In case some names have no corresponding item in the input parameter, they
will get \code{NULL} as value in the output list.
}
\examples{
\dontrun{
values <- extract(input, colClasses = c("date" = "Date", "text" = "character"))
}
}
