% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_view_ui.R
\name{filtered_view_ui}
\alias{filtered_view_ui}
\title{Filtered view}
\usage{
filtered_view_ui(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
a rendered DT data table
}
\description{
Filtered view
}
\examples{
filtered_view_ui(id = "mydata")
}
