% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_view_widget.R
\name{filtered_view_widget}
\alias{filtered_view_widget}
\title{Filtered View Widget}
\usage{
filtered_view_widget(id)
}
\arguments{
\item{id}{the id of the module server instance.}
}
\value{
An HTML object (tagList) to include in the UI.
}
\description{
Creates an HTML container for the filtered item table.
}
\examples{
\dontrun{
filtered_view_widget(id = "mydata")
}
}
