% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{Default Item Context}
\usage{
get_context()
}
\value{
a character value or NULL
}
\description{
Helper function to provide a contextual default value.
}
\details{
Since the project YAML config has been introduced (v0.8.0), all core functions
require to provide the name (id) of the item group to manipulate.

At the grammar level functions, this argument needs to be collected to pass
it's value to the core level functions.

When used inside the module server (one instance is dedicated to a specific item group)
or any custom code applied to a single item group, it would be tidious
to repeat this argument with the same value for each call.

By setting item = get_context(), the argument will get it's value from the
environment in which the function was called.
}
\examples{
# baseline
config <- design(project = "test", item = "foo")

# instead of
config |> extend(item = "foo", attribute = c(name = "total", type = "integer"))

# one can do
item <- "foo"
config |> extend(attribute = c(name = "total", type = "integer"))
}
\keyword{internal}
