% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_date_attribute.R
\name{has_date_attribute}
\alias{has_date_attribute}
\title{Has Date}
\usage{
has_date_attribute(config, item = get_context())
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}
}
\value{
a logical
}
\description{
Basic check if item has a 'date' attribute.
}
\details{
When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
has_date_attribute(design(project = "test"))
}
\seealso{
\code{\link[=parent.frame]{parent.frame()}}
}
