% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasDate.R
\name{hasDate}
\alias{hasDate}
\title{Test Date Attribute}
\usage{
hasDate(data.model)
}
\arguments{
\item{data.model}{a data.frame containing the data model.}
}
\value{
A logical, \code{TRUE} if \code{data.model} has an attribute "date" (exact match).
}
\description{
Test Date Attribute
}
\examples{
\dontrun{
hasDate(data.model = mydatamodel)
}
}
