% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_server.R
\name{import_server}
\alias{import_server}
\title{Import Module Server}
\usage{
import_server(id, k_data_model, k_items, callback)
}
\arguments{
\item{id}{the id of the module.}

\item{k_data_model}{the reference of the data model reactive.}

\item{k_items}{the reference of the items reactive.}

\item{callback}{a reactive to call once import is finished.}
}
\description{
The import module server is used as a child of the admin module server.
}
\keyword{internal}
