% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert Item(s)}
\usage{
insert(values, items)
}
\arguments{
\item{values}{a data.frame of item(s) to insert.}

\item{items}{a data.frame of items where to insert.}
}
\value{
a data.frame.
}
\description{
Insert Item(s)
}
\details{
The function is mostly a wrapper around \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} function.

It does check for potential duplicates among the 'id' column.
If so, the operation will be aborted as there is no way to know how to
fix the duplicated id(s).
}
\examples{
\dontrun{
insert(values, items)}
}
