% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_attribute}
\alias{is_attribute}
\title{Attribute Exists}
\usage{
is_attribute(config, item, attribute)
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}

\item{attribute}{a character vector holding the name of the attribute(s) to check}
}
\value{
a logical
}
\description{
Checks whether or not an attribute exists within an item.
}
\examples{
# build config (id attribute is mandatory)
config <- design(project = "test",
item = "foo")

# check
config |> is_attribute(item = "foo", attribute = "id")
config |> is_attribute(item = "foo", attribute = "dummy")
}
