% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_item}
\alias{is_item}
\title{Item Exists}
\usage{
is_item(config, item)
}
\arguments{
\item{config}{the config list}

\item{item}{the name (id) of the item group}
}
\value{
a logical
}
\description{
Check whether or not an item exists in the config.
}
\examples{
# build config
config <- design(project = "test", item = "foo")

# check
config |> is_item("foo")
config |> is_item("bar")
}
