% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_check.R
\name{item_check}
\alias{item_check}
\title{Check Items}
\usage{
item_check(items, config, id)
}
\arguments{
\item{items}{the data.frame of the items}

\item{config}{the YAML config list}

\item{id}{the name of the item}
}
\value{
a list
}
\description{
Check Items
}
\details{
It is expected that config is checked first using \code{\link[=config_check]{config_check()}}.
}
\examples{
\dontrun{
item_check(items, config, "foo")
}
}
\keyword{internal}
