% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_chk_str.R
\name{item_chk_str}
\alias{item_chk_str}
\title{Check Item Structure}
\usage{
item_chk_str(items, item)
}
\arguments{
\item{items}{the data.frame of the items}

\item{item}{the item to be tested}
}
\value{
either TRUE or an error
}
\description{
Check Item Structure
}
\details{
The function does not check the item vs the data model as it is used internally
during item creation or update (data model is not available in this context)
}
\examples{
\dontrun{
item_chk_str(items, item)
}
}
