% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create.R
\name{item_create}
\alias{item_create}
\title{Create item(s)}
\usage{
item_create(values, data.model)
}
\arguments{
\item{values}{a named list of values, matching with the data.model}

\item{data.model}{a data.frame, defining the data model}
}
\value{
a data.frame of the new item
}
\description{
Create item(s)
}
\details{
Names in values must fit with the data.model - no attribute should
be missing (extra ones would be ignored).

Case single item
Values should be submitted as a list of elements with length 1 or 0.
This means NULL, or length(0) values are accepted.

Case multiple items
Values can be a data.frame or a list of elements with length > 1
The provided list must be convertible into a data.frame

It is strongly advised to wrap calls to this function inside tryCatch
expressions as it will throw error rather than return a corrupted item.
}
\examples{
\dontrun{
item_create(values, data.model = mydatamodel)
}
}
