% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create_modal.R
\name{item_create_modal}
\alias{item_create_modal}
\title{Item Dialog}
\usage{
item_create_modal(data.model, items, shortcut, ns)
}
\arguments{
\item{data.model}{a data.frame with the data model information}

\item{items}{a data.frame with the items}

\item{shortcut}{a logical if shortcuts are activated or not}

\item{ns}{a namespace function}
}
\description{
Item Dialog
}
\examples{
\dontrun{
item_create_modal(data.model, items, shortcut, ns)
}
}
