% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create_workflow.R
\name{item_create_workflow}
\alias{item_create_workflow}
\title{Item Create Workflow}
\usage{
item_create_workflow(items, data.model, values)
}
\arguments{
\item{items}{a data.frame of the existing items}

\item{data.model}{a data.frame of the data model}

\item{values}{a list or data.frame of values to create new item(s)}
}
\value{
the new data.frame of the items
}
\description{
Item Create Workflow
}
\examples{
\dontrun{
item_create_workflow(items, data.model, values)
}
}
