% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_dialog.R
\name{item_dialog}
\alias{item_dialog}
\title{Item Modal Dialog(s)}
\usage{
item_dialog(..., workflow = c("create", "update", "delete"), ns)
}
\arguments{
\item{...}{the content to be displayed in the modal dialog.}

\item{workflow}{a character string to indicate workflow (see details).}

\item{ns}{the namespace function to use in the dialog.}
}
\value{
a modal dialog.
}
\description{
Produces a create / update / delete modal dialog.
}
\details{
Possible values for workflow are "create" (default), "update" or "delete".

\code{...} is typically the output of the item_form() function
When \code{workflow = "delete"}, it will be ignored and replaced by a standard message.
}
\examples{
\dontrun{
item_dialog(workflow = "delete", ns)
}
}
