% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_reveal.R
\name{item_reveal}
\alias{item_reveal}
\title{Reveal Items}
\usage{
item_reveal(items, data.model)
}
\arguments{
\item{items}{a data.frame of the items.}

\item{data.model}{a data.frame of the data model.}
}
\value{
A data.frame of the items with applied mask.
}
\description{
Apply data model display mask on the items.
}
\details{
The data model display mask is defined by attributes with display = \code{TRUE}.
}
\examples{
\dontrun{
item_reveal(items = "myitems", data.model = "mydatamodel")
}
}
