% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_search.R
\name{item_search}
\alias{item_search}
\title{Search Items}
\usage{
item_search(items, pattern)
}
\arguments{
\item{items}{a data.frame of items}

\item{pattern}{the search pattern}
}
\value{
a filtered data.frame
}
\description{
A simple wrapper around \link[dplyr]{filter} function
}
\details{
The function is not intended to perform any smart or advanced search. \cr
It provides basic search across all except \code{id} attributes.

Basically it returns \code{filter(items, if_any(-(id), ~ grepl(pattern, .)))}
}
\examples{
\dontrun{
item_search(items = mydata$items(), pattern = "Banana")
item_search(items = mydata$items(), pattern = 25)
}
}
