% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_update_workflow.R
\name{item_update_workflow}
\alias{item_update_workflow}
\title{Item Update Workflow}
\usage{
item_update_workflow(values, items, data.model)
}
\arguments{
\item{values}{a list or data.frame of values to create new item(s)}

\item{items}{a data.frame of the existing items}

\item{data.model}{a data.frame of the data model}
}
\value{
the new data.frame of the items
}
\description{
Item Update Workflow
}
\examples{
\dontrun{
item_update_workflow(items, data.model, values)
}
}
