% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_widget.R
\name{item_widget}
\alias{item_widget}
\title{Item Widget}
\usage{
item_widget(id)
}
\arguments{
\item{id}{the id of the module server instance.}
}
\value{
An HTML tag.
}
\description{
The UI component for the item table.
}
\examples{
\dontrun{
item_widget(id = "mydata")
}
}
