% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_tables_ui.R
\name{items_filtered_view_DT}
\alias{items_filtered_view_DT}
\title{Filtered view}
\usage{
items_filtered_view_DT(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
a rendered DT data table
}
\description{
Filtered view
}
\examples{
items_filtered_view_DT(id = "mydata")
}
