% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{items_name}
\alias{items_name}
\title{Items name}
\usage{
items_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Items name
}
\examples{
items_name(id = "mydata")
}
