% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{items_url}
\alias{items_url}
\title{Items URL}
\usage{
items_url(id, path = Sys.getenv("R_KITEMS_PATH"))
}
\arguments{
\item{id}{the id of the module server instance.}

\item{path}{the path to the data model and items.}
}
\value{
The URL of the items
}
\description{
Items URL
}
\details{
The recommended way to define the \code{path} argument is to set the R_KITEMS_PATH
environment variable.
}
\examples{
items_url(id = "mydata")
}
