% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems.R
\name{kitems}
\alias{kitems}
\title{Kitems Module Server}
\usage{
kitems(
  id,
  path = Sys.getenv("R_KITEMS_PATH"),
  trigger = NULL,
  filter = NULL,
  options = list(autosave = TRUE, notify = TRUE)
)
}
\arguments{
\item{id}{the unique id of the module server instance.}

\item{path}{where the YALM config and items are stored (see details).}

\item{trigger}{an optional reactive object to pass workflow events (see details).}

\item{filter}{an optional reactive object to pass filter events (see details).}

\item{options}{an optional list of options (see details).}
}
\value{
a list

Details about the elements of this list:
\itemize{
\item id = the \code{id} of the module (same as the input)
\item items = the reference of the items reactive object
\item data_model = a list, describing the item's data model
\item filtered_items = the reference of the filtered items reactive object
\item selected_items = the reference of the selected items (ids) reactive object
\item clicked_column = the reference of the clicked column reactive object
\item filters = the reference of the reactive list with filter expressions.
}
}
\description{
This is the main component of the package.
}
\details{
Since version 0.8.0, the recommended way to define the \code{path} argument is
to set the \code{R_KITEMS_PATH} environment variable.
If directly passed as an argument value, a warning will be raised at the console.

Behavior of the module server can be tuned using a list of options:
\itemize{
\item \code{autosave} is a logical whether the item auto save should be activated or not (default = \code{TRUE})
\item \code{notify} is a logical if Shiny notifications should be displayed (default = \code{TRUE})
Partial lists are supported, with missing elements getting the default value.
}

If autosave option is \code{FALSE}, the \code{item_save()} function should be used
to make the data persistent.

Triggers are the way to send events for the module to execute dedicated actions.
\code{trigger} must be a reactive (or \code{NULL}, the default).
An event is defined as a named list (see examples) and may be created
using the \code{\link[=trigger_event]{trigger_event()}} function.
If \code{NULL}, the trigger manager will not be initialized.

\code{filter} is a reactive object reference to pass filter expression(s) to the module server
A filter is defined as a named list (see examples).
It may be created using the \code{\link[=filter_event]{filter_event()}} helper function.
If \code{NULL}, the filter manager will not be initialized.
}
\examples{
\dontrun{
# baseline:
# launch module server to handle item group named "foo"
kitems(id = "foo")

# read only:
kitems(id = "foo", options = list(autosave = FALSE))

# trigger:
# pass workflow events to the module server
trigger <- reactiveVal()
kitems(id = "foo", trigger = trigger)

# fire the create item dialog:
trigger(list(workflow = "create", type = "dialog"))

# fire an item creation (no dialog):
# assuming 'foo' item has a 'name' character attribute
trigger(list(workflow = "create", type = "task", values = list(name = "test")))

# filter:
# pass filter events to the module server
filter <- reactiveVal()
kitems(id = "foo", filter = filter)

# apply filter at the pre-filtering layer
# assuming item 'foo' has a 'total' numeric attribute
filter(list(layer = "pre", expr = total > 10))
}
}
\seealso{
\code{\link[=trigger_event]{trigger_event()}}, \code{\link[=filter_event]{filter_event()}}
}
