% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_server.R
\name{kitems_server}
\alias{kitems_server}
\title{Module server}
\usage{
kitems_server(id, path, create = TRUE, autosave = TRUE)
}
\arguments{
\item{id}{the id to be used for the module server instance}

\item{path}{a path where data model and items will be stored}

\item{create}{a logical whether the data file should be created or not if missing (default = TRUE)}

\item{autosave}{a logical whether the item auto save should be activated or not (default = TRUE)}
}
\description{
Module server
}
\details{
If autosave is FALSE, the item_save function should be used to make the data persistent
}
\examples{
\dontrun{
kitems_server(id = "mydata", path = "path/to/my/data",
              create = TRUE, autosave = TRUE)
}
}
