% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Apply Mask}
\usage{
mask(x)
}
\arguments{
\item{x}{a data.frame or named object}
}
\value{
A renamed data.frame or object.
}
\description{
Apply mask on names.
}
\examples{
\dontrun{
mask(x = dm)
}
}
