% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_args.R
\name{match_arg_t1}
\alias{match_arg_t1}
\title{Match Argument Type 1}
\usage{
match_arg_t1(x, colClasses, mode = "any", name = deparse(substitute(x)))
}
\arguments{
\item{x}{the argument to check}

\item{colClasses}{the colClasses argument passed to the builder function}

\item{mode}{a mode to pass to \code{is.vector()}}

\item{name}{the name of the argument}
}
\description{
This internal function is mostly used to check the arguments
of the \code{data_model()} function.
}
\examples{
\dontrun{
foo <- match_arg_t1(foo, colClasses, mode = "character")
}
}
\keyword{internal}
