% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_args.R
\name{match_arg_t2}
\alias{match_arg_t2}
\title{Match Argument Type 2}
\usage{
match_arg_t2(x, default = FALSE, advice = NULL, name = deparse(substitute(x)))
}
\arguments{
\item{x}{the argument to check}

\item{default}{the default value for the argument}

\item{advice}{an optional verb to use for unsupported scenarios}

\item{name}{the name of the argument}
}
\description{
This internal function is mostly used to check the arguments
of the \code{data_model()} function.
}
\examples{
\dontrun{
foo <- match_arg_t2(foo, default = TRUE, advice = "skip")
}
}
\keyword{internal}
