% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare Values}
\usage{
prepare(values, config, update = FALSE, item = get_context())
}
\arguments{
\item{values}{a named list of values (see details).}

\item{config}{the config list.}

\item{update}{a logical if values are used in the update workflow (default \code{FALSE}).}

\item{item}{optional (see details), the name (id) of the item group.}
}
\value{
a data.frame
}
\description{
Turn trigger values into tabular data.
}
\details{
\code{values} is a named list. The names are used to check the corresponding values
vs the data.model (class, default values if the provided ones are not valid).
The elements in the list must have either length one or same length as the id element.

When an element has length one but the id has several values, all items corresponding
to these ids will be updated with same value. To do so, values will be turned into
a data.frame using as.data.frame ; for this reason, it's strongly advised to wrap
the call into \code{tryCatch()} as this may fail.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
\dontrun{
prepare(values, config, item = "foo")
}
}
\seealso{
\code{\link[=parent.frame]{parent.frame()}}
}
