% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_values.R
\name{prepare_values}
\alias{prepare_values}
\title{Prepare Values}
\usage{
prepare_values(values, data.model, update = FALSE)
}
\arguments{
\item{values}{a list of values.}

\item{data.model}{the data.frame of the data model.}

\item{update}{a logical if values are used in the update workflow (default FALSE).}
}
\value{
a data.frame of values
}
\description{
Turn trigger input values into tabular data
}
\examples{
\dontrun{
prepare_values(values, data.model)
}
}
