% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\alias{freeze}
\alias{refreshed}
\alias{frozen}
\title{Refresh Behavior Grammar}
\usage{
refresh(config, item = get_context(), ...)

freeze(config, item = get_context(), ...)

refreshed(config, item = get_context())

frozen(config, item = get_context())
}
\arguments{
\item{config}{the config list.}

\item{item}{the name (id) of the item group.}

\item{...}{the name of the attribute(s) to manipulate.}
}
\value{
a list (setters) or a character vector (getters).
}
\description{
A set of verbs to manipulate the refreshed attribute(s) of an item.
}
\details{
\code{refresh()} and \code{freeze()} are setter functions that allow to add or remove
attributes from the refreshed ones.
\code{refreshed()} and \code{frozen()} are getter functions to quickly access the skipped
attributes that are refreshed or not during an item update.

Trying to refresh an attribute that is not skipped will be ignored
\emph{without} a warning.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
\dontrun{
# refresh attribute
config <- design(project = "test",
item = "foo") |>
extend(item = "foo", attribute = c(name = "date", type = "Date")) |>
skip(item = "foo", "date") |>
refresh(item = "foo", "date")

# refreshed attributes
config |> refreshed(item = "foo")

# freeze attribute
config |> freeze(item = "foo", "date")

# frozen attributes
config |> frozen(item = "foo")
}
}
\seealso{
\code{\link[=amend]{amend()}}, \code{\link[=parent.frame]{parent.frame()}}
}
