% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_delete.R
\name{rows_delete}
\alias{rows_delete}
\title{Delete Item(s)}
\usage{
rows_delete(items, id)
}
\arguments{
\item{items}{the data.frame of the items.}

\item{id}{a vector of id(s) for the item(s) to delete.}
}
\value{
The updated data.frame of the items.
}
\description{
Delete Item(s)
}
\examples{
\dontrun{
rows_delete(items = myitems, id = 123456789)
}
}
