% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_insert.R
\name{rows_insert}
\alias{rows_insert}
\title{Insert Item(s)}
\usage{
rows_insert(values, items)
}
\arguments{
\item{values}{a data.frame of item(s) to insert.}

\item{items}{a data.frame of items where to insert.}
}
\value{
a data.frame of the items
}
\description{
Insert Item(s)
}
\examples{
\dontrun{
rows_insert(items, values)}
}
