% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run Shiny Demo Apps}
\usage{
runExample(example = NA)
}
\arguments{
\item{example}{a character value, with the name of the demo to be launched}
}
\description{
Run Shiny Demo Apps
}
\details{
When \code{example} is NA, the function will return the names of available apps
}
\examples{

\dontrun{
runExample()
}
}
