% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{selected_items_name}
\alias{selected_items_name}
\title{Selected items name}
\usage{
selected_items_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Selected items name
}
\examples{
selected_items_name(id = "mydata")
}
