% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\alias{include}
\alias{skipped}
\alias{included}
\title{Skip Behavior Grammar}
\usage{
skip(config, item = get_context(), ...)

include(config, item = get_context(), ...)

skipped(config, item = get_context())

included(config, item = get_context())
}
\arguments{
\item{config}{the config list.}

\item{item}{the name (id) of the item group.}

\item{...}{the name of the attribute(s) to manipulate.}
}
\value{
a list (setters) or a character vector (getters).
}
\description{
A set of verbs to manipulate the skipped attribute(s) of an item.
}
\details{
\code{skip()} and \code{include()} are setter functions that allow to add or remove
attributes from the skipped ones.
\code{skipped()} and \code{included()} are getter functions to quickly access the attributes
that are skipped or included in the item form.

In a scenario where items are created programmatically, \code{included()} may be used
to determine what values can be sent to the trigger.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
\dontrun{
# skip attribute
config <- design(project = "test",
item = "foo") |>
extend(item = "foo", attribute = c(name = "date", type = "Date")) |>
skip(item = "foo", "date")

# get skipped attributes
config |> skipped(item = "foo")

# include attribute
config |> include(item = "foo", "date")

# get included attributes
config |> included(item = "foo")
}
}
\seealso{
\code{\link[=amend]{amend()}}, \code{\link[=parent.frame]{parent.frame()}}
}
