% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{trigger_create_name}
\alias{trigger_create_name}
\title{Trigger create name}
\usage{
trigger_create_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Trigger create name
}
\examples{
trigger_create_name(id = "mydata")
}
