% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{trigger_delete_name}
\alias{trigger_delete_name}
\title{Trigger delete name}
\usage{
trigger_delete_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Trigger delete name
}
\examples{
trigger_delete_name(id = "mydata")
}
