% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_event.R
\name{trigger_event}
\alias{trigger_event}
\title{Trigger Event}
\usage{
trigger_event(
  workflow = c("create", "update", "delete"),
  type = c("dialog", "task"),
  values = NULL
)
}
\arguments{
\item{workflow}{the name of the workflow (create, update or delete).}

\item{type}{the name of the action to be performed (dialog or task).}

\item{values}{optional list of values to create, update or delete an item.}
}
\value{
A list.
}
\description{
Helper function to create events to be passed to kitems module server
with the \code{trigger} argument.
}
\details{
The function also adds an event id to make it unique (otherwise sending
two times the same request would fail).

For more details see this vignette:
\code{vignette("workflows", package = "kitems")}
}
\examples{
\dontrun{
# create dialog event
trigger_event()

# delete item event (without dialog)
trigger_event(workflow = "delete", type = "task", values = list(id = 1234))
}
}
\seealso{
\code{\link[=kitems]{kitems()}}
}
