% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update Item(s)}
\usage{
update(values, items)
}
\arguments{
\item{values}{a data.frame used to update the items.}

\item{items}{a data.frame of the items.}
}
\value{
a data.frame.
}
\description{
Update Item(s)
}
\details{
The function will drop unmatched id(s) in \code{values}.
}
\examples{
\dontrun{
update(values, items)}
}
