% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons_ui.R
\name{update_BTN}
\alias{update_BTN}
\title{Update item button}
\usage{
update_BTN(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a rendered actionButton
}
\description{
Update item button
}
\examples{
update_BTN(id = "mydata")
}
