% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate Value(s)}
\usage{
validate(values, data.model, items = NULL, update = FALSE)
}
\arguments{
\item{values}{a list of named values.}

\item{data.model}{the data.frame of the data model (see details).}

\item{items}{an optional data.frame of the items (see details).}

\item{update}{whether the id attribute should be checked or not (default \code{FALSE}).}
}
\value{
A data.frame.
}
\description{
Validate values and turn them into item(s).
}
\details{
This function does not accept the data.model element of the config list as an input.
It should be turned into a data.frame first, using \code{\link[=yaml_to_dm]{yaml_to_dm()}}.

\code{data.model} should contain the following columns:
"name", "type", "default", "class.arg", "values".

In case the data.model values column contains instruction that require
data-masking, \code{items} will be used as a context to evaluate those instructions.
}
\examples{
\dontrun{
# create config
config <- design(project = "test", item = "foo") |>
extend(item = "foo",
attribute = list(name = "name", type = "character"),
attribute = list(name = "quantity", type = "numeric"))

# values
values <- list(name = c("Banana", "Mango"), quantity = c(1, 12))

# call function
validate(values, data.model = yaml_to_dm(config, "name", "type", "default", "class.arg", "values"))
}
}
