% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_widgets.R
\name{actions_widget}
\alias{actions_widget}
\title{Buttons}
\usage{
actions_widget(id)
}
\arguments{
\item{id}{the server module id.}
}
\value{
a list of HTML tags.
}
\description{
Buttons to fire the create / update / delete dialog.
}
\examples{
actions_widget("foo")
}
