% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Sort Items}
\usage{
adjust(items, config, item = get_context())
}
\arguments{
\item{items}{a data.frame of the items.}

\item{config}{the config list.}

\item{item}{optional (see details), the name (id) of the item group.}
}
\value{
a data.frame.
}
\description{
Sort Items
}
\details{
The sorting order is given by the sort entry in the config list.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
\dontrun{
items |> adjust(config, item = "foo")
}
}
\seealso{
\link[=parent.frame]{organize()}
}
