% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin}
\alias{admin}
\title{Admin Console}
\usage{
admin()
}
\description{
Launches the administration console (Shiny App)
}
\examples{
\dontrun{
# set environment (where to find the _kitems.yml)
Sys.setenv("R_KITEMS_PATH" = "D:/data")

# launch the Admin Console
admin()
}
}
