% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin}
\alias{admin}
\title{Admin Console}
\usage{
admin(path = getwd())
}
\arguments{
\item{path}{the path where to find the item folder(s)}
}
\description{
Launches the administration console (Shiny App)
}
\details{
The app will scan the \code{path} to detect sub folders that are expected to be
item folders named after the id used to create them.

It will build the ui tabs from this list.
}
\examples{
\dontrun{
admin()
}
}
