% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_attribute_layouts.R
\name{admin_attribute_card}
\alias{admin_attribute_card}
\title{Attribute Card}
\usage{
admin_attribute_card(attribute, item, hide = NULL, skip = NULL, refresh = NULL)
}
\arguments{
\item{attribute}{the yaml of the attribute}

\item{item}{the name of the item}

\item{hide}{the list of attributes to hide}

\item{skip}{the list of attributes to skip}

\item{refresh}{the list of attributes to refresh}
}
\value{
A htmltools::div() tag.
}
\description{
This is a UI layout function to display attributes.
}
\examples{
\dontrun{
admin_attribute_card(attribute, item)
}
}
