% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_attribute_layouts.R
\name{admin_attribute_modal}
\alias{admin_attribute_modal}
\title{Attribute Modal Window}
\usage{
admin_attribute_modal(update = FALSE)
}
\arguments{
\item{update}{whether it is an update or not}
}
\value{
a modal dialog
}
\description{
This is the UI layout function for the attribute wizard.
}
\details{
Mostly composed of container div elements that will be filled / updated
dynamically.
}
\examples{
\dontrun{
admin_attribute_modal()
}
}
