% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_layouts.R
\name{admin_attribute_nb}
\alias{admin_attribute_nb}
\title{Attribute Number}
\usage{
admin_attribute_nb(config)
}
\arguments{
\item{config}{an item config}
}
\value{
a character string
}
\description{
Attribute Number
}
\examples{
\dontrun{
admin_attribute_nb(config)
}
}
