% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_attribute_wizard.R
\name{admin_attribute_wizard}
\alias{admin_attribute_wizard}
\title{Attribute Wizard}
\usage{
admin_attribute_wizard(
  config,
  item,
  attribute = NULL,
  callback,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{config}{the YAML config}

\item{item}{the name (id) of the item group}

\item{attribute}{an optional name to indicate it's an update}

\item{callback}{a reactiveVal to send the output of the wizard}

\item{session}{optional, the Shiny session object}
}
\description{
The function launches a modal window to display a step by step
attribute wizard.
}
\examples{
\dontrun{
admin_attribute_wizard(config, item = "foo")
}
}
