% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_item_layouts.R
\name{admin_item_card}
\alias{admin_item_card}
\title{Item Card}
\usage{
admin_item_card(name, description = NULL)
}
\arguments{
\item{name}{the name of the item}

\item{description}{an optional description}
}
\value{
a div() HTML object
}
\description{
This is a UI / Layout function for the Admin Console
}
\examples{
\dontrun{
admin_item_card("date")
}
}
