% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_item_layouts.R
\name{admin_item_dz}
\alias{admin_item_dz}
\title{Item Danger Zone}
\usage{
admin_item_dz(name)
}
\arguments{
\item{name}{the name of the item}
}
\value{
an HTML tag
}
\description{
This is a UI / Layout function for the Admin Console
}
\examples{
\dontrun{
admin_item_dz("date")
}
}
