% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_migration_layouts.R
\name{admin_migration_done_layout}
\alias{admin_migration_done_layout}
\title{Migration Done Layout}
\usage{
admin_migration_done_layout(path)
}
\arguments{
\item{path}{the path of the project}
}
\value{
an HTML layout
}
\description{
Migration Done Layout
}
\examples{
\dontrun{
admin_migration_done_layout(path = ".")
}
}
