% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_layouts.R
\name{admin_no_yaml_layout}
\alias{admin_no_yaml_layout}
\title{Missing Config Layout}
\usage{
admin_no_yaml_layout()
}
\value{
an HTML tag (div)
}
\description{
This is a UI / Layout function for the Admin Console
}
\examples{
\dontrun{
admin_no_yaml_layout()
}
}
