% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_widget.R
\name{admin_widget}
\alias{admin_widget}
\title{Admin Widget}
\usage{
admin_widget(id)
}
\arguments{
\item{id}{the module id}
}
\value{
a tagList() object to admin the data model
}
\description{
Admin Widget
}
\examples{
admin_widget(id = "mydata")
}
