% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend.R
\name{amend}
\alias{amend}
\title{Amend Attribute}
\usage{
amend(config, item = get_context(), ...)
}
\arguments{
\item{config}{the config list.}

\item{item}{optional (see details), the name (id) of the item group.}

\item{...}{one or several attribute instructions.}
}
\value{
a config list
}
\description{
Update specific parameter(s) of specific attribute(s)).
}
\details{
The function only understands \code{attribute} instruction.
It can handle multiple instructions in \code{...}.

Note that it is forbidden to update:
\itemize{
\item the id attribute (any parameter),
\item the name and/or type of an attribute.
}

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
# create baseline
config <- design(project = "test",
                 item = "foo") |>
  extend(item = "foo",
         attribute = c(name = "total", type = "integer"))

# skip item argument (in the following calls)
item <- "foo"

# single instruction
config |>
  amend(attribute = c(name = "total", default = 12))
config |>
  amend(attribute = c(name = "total", values = "suggest(12)"))

# multiple instructions
config |>
  amend(attribute = c(name = "total", values = "suggest(12)"),
        attribute = c(name = "total", default = "0"))
}
\seealso{
\code{\link[=parent.frame]{parent.frame()}}
}
