% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_default.R
\name{as_default}
\alias{as_default}
\title{Turn Item Into Default Value(s)}
\usage{
as_default(item, data.model)
}
\arguments{
\item{item}{the data.frame of the item to use as a reference.}

\item{data.model}{the data.frame of the data.model (see details).}
}
\value{
a data.frame to pass to form() function.
}
\description{
Turn Item Into Default Value(s)
}
\details{
This function does not accept the data.model element of the config list as an input.
It should be turned into a data.frame first, using \code{\link[=yaml_to_dm]{yaml_to_dm()}}.

\code{data.model} should contain the following columns:
"name", "type", "default", "values".

\code{item} and \code{data.model} must have same structure. That means the names
of the attributes in the data model are expected to match with the names
of the columns in item.
}
\examples{
\dontrun{
as_default(item, data.model)
}
}
