% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_default.R
\name{as_default}
\alias{as_default}
\title{Turn Item Into Default Value(s)}
\usage{
as_default(item, data.model)
}
\arguments{
\item{item}{the data.frame of the item to use as a reference.}

\item{data.model}{the data.frame of the data.model.}
}
\value{
a data.frame to pass to item_form() function.
}
\description{
Turn Item Into Default Value(s)
}
\details{
\code{item} and \code{data.model} must have same structure. That means the names
of the attributes in the data model are expected to match with the names
of the columns in item.
}
\examples{
\dontrun{
as_default(item, data.model)
}
}
