% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_create.R
\name{attribute_create}
\alias{attribute_create}
\title{Add Attribute}
\usage{
attribute_create(
  data.model = NULL,
  name,
  type,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  display = NULL,
  skip = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{data.model}{a data model data.frame, structured as an output of the \code{data_model()} function}

\item{name}{a character vector for the new attribute name}

\item{type}{a character vector for the new attribute type}

\item{default.val}{an optional named vector of values, defining the default values.}

\item{default.fun}{an optional named vector of functions, defining the default functions to be used to generate default values.}

\item{default.arg}{an optional named vector of arguments, to pass along with the default function.}

\item{display}{an optional character vector, with the name(s) of the attribute(s) to display}

\item{skip}{an optional character vector, with the name(s) of the attribute(s) to skip}

\item{sort.rank}{an optional named numeric vector, to define sort orders}

\item{sort.desc}{an optional named logical vector, to define if sort should be descending}
}
\value{
The updated data model data.frame
}
\description{
Add an attribute to a data model
}
\details{
Multiple attribute creation is supported, in this case make sure all vectors have same length.

When \code{data.model} is omitted, the function will return a data model containing the created attributes.
}
\examples{
\dontrun{

# -- create single attribute
attribute_create(name = "new_attribute", type = "character")
attribute_create(name = "total", type = "numeric", default.val = 0)
attribute_create(name = "date", type = "Date", default.fun = "Sys.Date")
attribute_create(name = "progress", type = "integer", skip = "progress")
attribute_create(name = "internal", type = "logical", display = "internal")

# -- create multiple attributes
attribute_create(name = c("foo", "bar"), type = c("character", "numeric"))

}
}
\seealso{
\code{\link[=data_model]{data_model()}}
}
