% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_delete.R
\name{attribute_delete}
\alias{attribute_delete}
\title{Delete attribute}
\usage{
attribute_delete(
  k_data_model,
  k_items,
  name,
  MODULE = NULL,
  autosave = FALSE,
  dm_url = NULL,
  items_url = NULL,
  notify = FALSE
)
}
\arguments{
\item{k_data_model}{the reference of the data model reactive value}

\item{k_items}{the reference of the items reactive value}

\item{name}{the name of the attribute to delete}

\item{MODULE}{an optional string to be displayed in the notificatlion}

\item{autosave}{a logical if autosave is ON}

\item{dm_url}{the url of the data model file}

\item{items_url}{the url of the item file}

\item{notify}{a logical if shiny notificatlion should be fired}
}
\description{
Delete attribute
}
\examples{
\dontrun{
attribute_delete(k_data_model,
k_items,
name = "comment",
MODULE = "mydata",
autosave = TRUE,
dm_url = dm_url,
items_url = items_url,
notify = TRUE)
}

}
