% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_input.R
\name{attribute_input}
\alias{attribute_input}
\title{Attribute Input}
\usage{
attribute_input(name, type, value = NULL, ns)
}
\arguments{
\item{name}{the name of the attribute.}

\item{type}{the type of the attribute.}

\item{value}{the value to be used to initialize the input.}

\item{ns}{the namespace function reference.}
}
\value{
An input that can be added to the UI definition.
}
\description{
Build input for an attribute
}
\examples{
\dontrun{
# -- namespace
ns <- shiny::NS("my_data")

# -- create inputs
attribute_input(name = "total", type = "numeric", value = 10, ns)
}
}
