% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_input.R
\name{attribute_input}
\alias{attribute_input}
\title{Build Attribute Input}
\usage{
attribute_input(colClass, value = NULL, ns)
}
\arguments{
\item{colClass}{a length-one named vector. \code{names(colClass)} is the name of the attribute,
and \code{colClass} is the type (class) of the attribute.}

\item{value}{the value used to initialize the input}

\item{ns}{the module namespace function reference}
}
\value{
an input based on colClass
}
\description{
Build Attribute Input
}
\examples{
\dontrun{
attribute_input(colClass = c(name = "character"), ns)
attribute_input(colClass = c(total = "numeric"), value = 10, ns)
}
}
