% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_input_update.R
\name{attribute_input_update}
\alias{attribute_input_update}
\title{Update Attribute Input}
\usage{
attribute_input_update(data.model, xxxx_trigger, MODULE = NULL)
}
\arguments{
\item{data.model}{the reference of the data model reactiveVal}

\item{xxxx_trigger}{basically the value of input$xxxx_trigger}

\item{MODULE}{an optional character string for the trace}
}
\description{
Update Attribute Input
}
\examples{
\dontrun{
attribute_input_update(data.model = mydata$data_model, xxxx_trigger = "name_banana", MODULE = "(mydata)")
}
}
