% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_suggestion.R
\name{attribute_suggestion}
\alias{attribute_suggestion}
\title{Attribute Suggestions}
\usage{
attribute_suggestion(values, type = class(values), n = 3, floor = 10)
}
\arguments{
\item{values}{a vector of values, most probably a vector corresponding to one of the item data frame columns}

\item{type}{an optional character vector, the type (class) of values. If not provided, class(values) will be used}

\item{n}{the desired number of suggestions. This applies only on character, factor, numeric & integer types}

\item{floor}{a numeric value. This applies only for numeric & integer types. If none of the value frequency (in \%)
reaches \code{floor}, then basic suggestions are returned (see details)}
}
\value{
a list of suggestions
This list depends on the \code{type} of values
\itemize{
\item character, factor:
the most frequent values ; \code{list(value1 = frequency, value_2, = frequency)}
\item numeric, integer:
the most frequent values ; \code{list(value1 = frequency, value_2, = frequency)}, or
a standard list ; \code{list(min, max, mean, median)}
\item logical:
the frequency for both values ; \code{list(true = frequency, false = frequency)}
\item Date, POSIXct:
a standard list ; \code{list(min, max)}
}
}
\description{
Attribute Suggestions
}
\details{
For numeric & integer values, the default strategy is to compute the \code{n} most frequent occurrences.
If no occurrence reaches the \code{floor} level (\% of all items), then the standard list is returned.
}
\examples{
foo <- c(rep("banana", 5), rep("mango", 3), rep("orange", 2))
attribute_suggestion(values = foo, type = "character")
}
