% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_update.R
\name{attribute_update}
\alias{attribute_update}
\title{Update Attribute}
\usage{
attribute_update(
  data.model,
  name,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  display = NULL,
  skip = NULL,
  refresh = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{name}{a character string of the attribute name}

\item{default.val}{a character string, the new default value}

\item{default.fun}{a character string, the new default function name}

\item{default.arg}{a named vector of arguments, to pass along with the default function}

\item{display}{a logical to set the display value}

\item{skip}{a logical to set the skip value}

\item{refresh}{a logical to tell if skipped attributes should be refreshed upon update}

\item{sort.rank}{a numeric, used to define sort rank}

\item{sort.desc}{a logical, to define if sorting should be in descending order}
}
\value{
The updated data model data.frame
}
\description{
Update a data model attribute
}
\details{
Updating attribute class is not supported by this function (as it requires data migration).

The use of vectors to update several attributes is supported as long as the \code{length()} of the different
parameters is either same as \code{name} or 1 (then all rows gets same value).
}
\examples{

# -- define dm
dm <- data_model(colClasses = c(name = "character", total = "numeric"))

# -- Use of vectors to update several attributes
attribute_update(data.model = dm,
                 name = c("name","total"),
                 default.val = c("test", 2),
                 skip = TRUE)

}
\seealso{
\code{\link[=data_model]{data_model()}}
}
