% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_update.R
\name{attribute_update}
\alias{attribute_update}
\title{Update Attribute}
\usage{
attribute_update(
  data.model,
  name,
  class.arg = NULL,
  default.val = NULL,
  default.fun = NULL,
  display = NULL,
  skip = NULL,
  refresh = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{name}{the name(s) of the attribute(s) to update (see details)}

\item{class.arg}{the argument(s) to pass along with the class function}

\item{default.val}{the new default value(s)}

\item{default.fun}{the new default function name(s)}

\item{display}{a logical to set display}

\item{skip}{a logical to set skip}

\item{refresh}{a logical to tell if skipped attributes should be refreshed upon update}

\item{sort.rank}{a named vector of integer(s), used to define sort rank}

\item{sort.desc}{a named vector of logical(s), to define if sorting should be in descending order}
}
\value{
The updated data.model
}
\description{
Update data model attribute(s)
}
\details{
Updating attribute class is not supported (as it requires data migration).

Partial update will be performed based on arguments that are not NULL.

The use of vectors to update several attributes is supported as long as the \code{length()} of the different
parameters is either same as \code{name} or 1 (then all rows gets same value).
When the intend is to update all attributes with the same value, then named vector is not necessary.

When attribute parameters are not compatible, they will be reset.
Ex: setting \code{default.val} will reset \code{default.fun}

Otherwise same rules apply as in the \code{data_model()} function.
}
\examples{

# -- define dm
dm <- data_model(colClasses = c(name = "character", total = "numeric"))

# -- Use of vectors to update several attributes
attribute_update(data.model = dm,
                 name = c("name", "total"),
                 default.val = c(name = "test", total = 2),
                 skip = TRUE)

}
\seealso{
\code{\link[=data_model]{data_model()}}
}
