% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_update.R
\name{attribute_update}
\alias{attribute_update}
\title{Update data model attribute}
\usage{
attribute_update(
  data.model,
  name,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  skip = NULL,
  display = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{name}{a character string of the attribute name}

\item{default.val}{a character string, the new default value}

\item{default.fun}{a character string, the new default function name}

\item{default.arg}{an optional named vector of arguments, to pass along with the default function.}

\item{skip}{a logical to set the skip value}

\item{display}{a logical to set the display value}

\item{sort.rank}{a numeric, used to define sort rank}

\item{sort.desc}{a logical, to define if sorting should be in descending order}
}
\value{
the updated data model
}
\description{
Update data model attribute
}
\details{
Updating attribute class is not supported by this function (as it requires data migration).

Use of vector to update several attributes is supported as long as length of the different parameters
is either same as name or 1 (then all rows gets same value).
}
\examples{
\dontrun{
#Use of vector to update several attributes:
attribute_update(data.model = dm,
                 name = c("name","total"),
                 default.val = c("test", 2),
                 default.fun = NULL,
                 default.arg = NULL,
                 skip = TRUE)
}
}
\seealso{
\code{\link[=data_model]{data_model()}}
}
