% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_values.R
\name{attribute_values}
\alias{attribute_values}
\title{Validate Value(s)}
\usage{
attribute_values(values, data.model, update = FALSE)
}
\arguments{
\item{data.model}{the data.frame of the data model.}

\item{update}{whether the id attribute should be checked or not (default FALSE).}

\item{value}{a list of input values.}
}
\value{
A data.frame of item(s) checked against the data model.
}
\description{
Validate the input values and turn them into item(s)
}
\examples{
\dontrun{
attribute_values(values, data.model)}
}
