% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_values.R
\name{attribute_values}
\alias{attribute_values}
\title{Turn Value(s) into Attribute}
\usage{
attribute_values(key, value, data.model)
}
\arguments{
\item{key}{the name of the attribute.}

\item{value}{a vector of input values.}

\item{data.model}{the data.frame of the data model.}
}
\value{
A vector of values checked against the data model.
}
\description{
Turn Value(s) into Attribute
}
\examples{
\dontrun{
attribute_values(key, value, data.model)}
}
