% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{Backup Files}
\usage{
backup(id, path = Sys.getenv("R_KITEMS_PATH"), type = "items", max = NULL)
}
\arguments{
\item{id}{the id used to create the data model.}

\item{path}{the path to the data model and items.}

\item{type}{the type of file to backup: "items" (default) or "dm".}

\item{max}{an integer to indicate how many backup files are allowed.}
}
\description{
Backup data model & items files
}
\details{
The recommended way to define the \code{path} argument is to set the R_KITEMS_PATH
environment variable.

Backup file will be named as \emph{id_data_model_YYYY-MM-DD.rds} or \emph{id_items_YYYY-MM-DD.csv}
If same file already exists, it will be overwritten.

If the number of backup files exceeds \code{max} then the oldest will be deleted.
Whenever \code{max = NULL} (default), it will be replaced by 1.
}
\examples{
\dontrun{
backup(id = "mydata", max = 2)
}
}
