% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{Backup Data Model File}
\usage{
backup(id, path, type = "items", max = NULL)
}
\arguments{
\item{id}{the kitems id used to create the data model}

\item{path}{the path to the data model}

\item{type}{the type of file to backup. \code{items} (default) or \code{dm}}

\item{max}{an integer to indicate how many backup files are allowed}
}
\description{
Backup Data Model File
}
\details{
Backup file will be named as \emph{id_data_model_YYYY-MM-DD.rds}
If same file already exists, it will be overwritten.

If the number of backup files exceeds \code{max} then the oldest will be deleted.
Whenever \code{max = NULL} (default), it will be replaced by 1
}
\examples{
backup(id = "mydata", path = "path/to/my/data", max = 2)
}
