% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check Integrity}
\usage{
check(...)
}
\arguments{
\item{...}{the object(s) to check (see details)}
}
\value{
a list
}
\description{
Wrapper function to either check the YAML config or items integrity.
}
\details{
The function is smart enough to detect what is passed to \code{...}.
To check the items, it is necessary to also pass the corresponding
id in the config (see examples).
}
\examples{
\dontrun{
# check config (list)
config <- design(project = "test")
check(config)

# check items (data.frame)
check(items, config, id = "foo")
}
}
