% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check Path Argument}
\usage{
check_path(path)
}
\arguments{
\item{path}{the value to check}
}
\value{
NULL if all checks are ok
}
\description{
Check Path Argument
}
\examples{
check_path("./")
}
