% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Attribute Values}
\usage{
compute(x, data = NULL)
}
\arguments{
\item{x}{a character string of the expression to evaluate}

\item{data}{a data.frame to use for tidy evaluation of \code{x}}
}
\value{
a vector of values
}
\description{
Compute attribute values with data-masking support
}
\examples{
compute(x = "suggest(1, 2)")
compute(x = "suggest(min(value))", data = data.frame(value = c(12, 10)))
}
