% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_append}
\alias{config_attribute_append}
\title{Append Attribute}
\usage{
config_attribute_append(config, item, ...)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}

\item{...}{one or several attribute config(s)}
}
\value{
the new config
}
\description{
Append Attribute
}
\examples{
\dontrun{
config_attribute_append(config, item = foo,
                        attribute = config_attribute_create(name = "bar", type "numeric"))
}
}
