% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attribute_create}
\alias{config_attribute_create}
\title{Create Attribute}
\usage{
config_attribute_create(
  name,
  type,
  class.arg = NULL,
  values = NULL,
  default = NULL
)
}
\arguments{
\item{name}{the name of the attribute}

\item{type}{the type of the attribute}

\item{class.arg}{optional argument to sent to the as.* conversion function}

\item{values}{a character string to set the rules for values}

\item{default}{a character string to define the defaults}
}
\value{
a list
}
\description{
Create Attribute
}
\examples{
\dontrun{
config_attribute_create(name = "att_1", type = "integer")
}
}
