% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_attributes}
\alias{config_attributes}
\title{Config Attributes}
\usage{
config_attributes(config, item)
}
\arguments{
\item{config}{a yaml config section for an item}

\item{item}{the name of the item}
}
\value{
a vector
}
\description{
Get the attribute names in an item YAML config section.
}
\examples{
\dontrun{
config_attributes(config)
}
}
