% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_move}
\alias{config_item_move}
\title{Move Item}
\usage{
config_item_move(config, item, where)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name (id) of the item group}

\item{where}{a list(position = c("before", "after"), item = "foo")}
}
\value{
the new config
}
\description{
Move Item
}
\examples{
\dontrun{
config_item_move(config, item = "bar", where = list(position = "before", item = "foo"))
}
}
