% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_item_sort}
\alias{config_item_sort}
\title{Sort Items}
\usage{
config_item_sort(config, item, sort = NULL)
}
\arguments{
\item{config}{the yaml config}

\item{item}{the name of the item to drop}

\item{sort}{a character string}
}
\value{
the new config
}
\description{
Sort Items
}
\details{
\code{sort} will be used to order the items.
Comma should be used between attribute names.
Wrap attribute name with desc() to indicate descending order.
}
\examples{
\dontrun{
# define sorting
config_item_sort(config, item, sort = "date, value")
config_item_sort(config, item, sort = "desc(date), value")

# reset sorting
config_item_sort(config, item, sort = NULL)
}
}
