% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_items}
\alias{config_items}
\title{Config Items}
\usage{
config_items(config)
}
\arguments{
\item{config}{a yaml config}
}
\value{
a vector
}
\description{
Get the item names (id) in a YAML config.
}
\examples{
config_items(list(items = list(config_item_create(id = "foo", path = "./"),
                               config_item_create(id = "bar", path = "./"))))
}
