% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_read.R
\name{config_read}
\alias{config_read}
\title{Read YAML config file}
\usage{
config_read(path = Sys.getenv("R_KITEMS_PATH"))
}
\arguments{
\item{path}{where to find _kitems.yml}
}
\value{
a list or NULL if no file is found.
}
\description{
Read YAML config file
}
\details{
By default, \code{path} uses the R_KITEMS_PATH environment variable.
}
\examples{
\dontrun{
config_read()
}
}
