% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_widgets.R
\name{create_widget}
\alias{create_widget}
\title{Create Item Button}
\usage{
create_widget(id)
}
\arguments{
\item{id}{the server module id.}
}
\value{
An HTML element that can be included in the UI.
}
\description{
Button to fire the create item dialog.
}
\examples{
\dontrun{
create_widget(id = "mydata")
}
}
