% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_widgets.R
\name{create_widget}
\alias{create_widget}
\title{Create Item Button}
\usage{
create_widget(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a rendered actionButton
}
\description{
Create Item Button
}
\examples{
create_widget(id = "mydata")
}
