% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{data_model}
\alias{data_model}
\title{Build data model}
\usage{
data_model(
  colClasses,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  display = NULL,
  skip = NULL,
  sort.rank = NULL,
  sort.desc = NULL
)
}
\arguments{
\item{colClasses}{a \emph{mandatory} named vector of classes, defining the data model.}

\item{default.val}{an optional named vector of values, defining the default values.}

\item{default.fun}{an optional named vector of functions, defining the default functions to be used to generate default values.}

\item{default.arg}{an optional named vector of arguments, to pass along with the default function.}

\item{display}{an optional character vector, indicating which attribute names should be displayed in the table view.}

\item{skip}{an optional character vector, indicating which attribute names should be skipped from the user input form.}

\item{sort.rank}{an optional named numeric vector, to define sort orders.}

\item{sort.desc}{an optional named logical vector, to define if sort should be descending.}
}
\value{
a data.frame containing the data model.
}
\description{
Build data model
}
\details{
colClasses will be used to create the data.frame: names will define the attributes of the data model,
and values will define the class (type) of the attributes.

All default.* parameters are optional. When provided, they will be used to match with names defined in colClasses:
\itemize{
\item order in those vectors doesn't matter
\item there is no need to set values for all attributes (see examples)
\item names in vectors not matching with colClasses names will be ignored
}

default.fun and default.val are mutual exclusive, with priority on default.fun (default.val will be forced to NA)
default.arg requires default.fun not to be NULL (will be forced to NA otherwise)

display and skip directly contains the names of the attributes to set to TRUE

If not provided, defaults will be applied:
\itemize{
\item NA for default.val, default.fun and default.arg
\item FALSE for display and skip
}
}
\examples{
# -- order in vectors doesn't matter:
default.val <- c("name" = "test", "total" = 2)
default.val <- c("total" = 2, "name" = "test")

# -- no need to set all values
colClasses <- c("id" = "numeric", "name" = "character", "total" = "numeric")
default.val <- c("name" = "test", "total" = 2)

# -- display and skip
display <- "id"
skip <- c("id", "date")

# -- sort
sort.rank = c("date" = 1, "total" = 2, "name" = 3)
sort.desc = c("date" = TRUE, "total" = FALSE)

data_model(colClasses, default.val, display = display, skip = skip)

}
