% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_slider_widget.R
\name{date_slider_widget}
\alias{date_slider_widget}
\title{Date Slider Widget}
\usage{
date_slider_widget(id, width = "300px", bootstrap = bslib::version_default())
}
\arguments{
\item{id}{the server module id.}

\item{width}{the width of the input (passed to \code{shiny::sliderInput()}).}

\item{bootstrap}{optional, a character string for the Bootstrap major version (see details).}
}
\value{
An HTML component to include on UI side.
}
\description{
The date slider UI component to drive the date filter.
}
\details{
\code{bootstrap} has been added in version 0.8.0 to determine if \code{bslib::popover()} can be used
in the widget. The default is provided by \code{bslib::version_default()}, but there is no obvious
way to know - from the UI side - if the app uses shiny's default or bslib layouts.

In case you are using shiny's layout (ex: \code{shiny::fluidPage()}), you should set the parameter
to \code{bootstrap = "3"}.
}
\examples{
\dontrun{
date_slider_widget(id = "mydata")
}
}
