% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_slider_widget.R
\name{date_slider_widget}
\alias{date_slider_widget}
\title{Date slider Widget}
\usage{
date_slider_widget(id)
}
\arguments{
\item{id}{the server module id}
}
\value{
a ui object (output of fluidRow)
}
\description{
Date slider Widget
}
\examples{
date_slider_widget(id = "mydata")
}
