% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_slider_widget.R
\name{date_slider_widget}
\alias{date_slider_widget}
\title{Date Slider Widget}
\usage{
date_slider_widget(id)
}
\arguments{
\item{id}{the server module id.}
}
\value{
an HTML component to include on UI side.
}
\description{
The date slider UI component to drive the date filter
}
\examples{
\dontrun{
date_slider_widget(id = "mydata")
}
}
