% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Decorate Item Columns}
\usage{
decorate(x)
}
\arguments{
\item{x}{a data.frame or named object}
}
\value{
A renamed data.frame or object.
}
\description{
Apply mask on attribute names.
}
\examples{
\dontrun{
decorate(x = dm)
}
}
