% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default}
\alias{default}
\title{Default Value(s)}
\usage{
default(data.model)
}
\arguments{
\item{data.model}{the data.frame of the data model.}
}
\value{
A data.frame
}
\description{
Compute default value(s).
}
\details{
This function does not accept the data.model element of the config list as an input.
It should be turned into a data.frame first, using \code{\link[=yaml_to_dm]{yaml_to_dm()}}.

\code{data.model} should contain the following columns:
"name", "type", "default" & "values".
}
\examples{
\dontrun{
default(yaml_to_dm(config...)) # << check this!
}
}
