% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete Item(s)}
\usage{
delete(items, id)
}
\arguments{
\item{items}{the data.frame of the items.}

\item{id}{a vector of id(s) for the item(s) to delete.}
}
\value{
The updated data.frame of the items.
}
\description{
Delete Item(s)
}
\examples{
\dontrun{
delete(items = myitems, id = 123456789)
}
}
