% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_widgets.R
\name{delete_widget}
\alias{delete_widget}
\title{Delete Item Button}
\usage{
delete_widget(id)
}
\arguments{
\item{id}{the server module id.}
}
\value{
An HTML element that can be included in the UI.
}
\description{
Button to fire the delete item dialog.
}
\examples{
\dontrun{
delete_widget(id = "mydata")
}
}
