% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_add_attribute.R
\name{dm_add_attribute}
\alias{dm_add_attribute}
\title{Title}
\usage{
dm_add_attribute(
  data.model,
  name,
  type,
  default.val = NA,
  default.fun = NA,
  skip = FALSE,
  filter = FALSE
)
}
\arguments{
\item{filter}{}
}
\description{
Title
}
