% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_colClasses.R
\name{dm_colClasses}
\alias{dm_colClasses}
\title{Data Model colClasses}
\usage{
dm_colClasses(data.model = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model.}
}
\value{
A named vector of the attribute types.
}
\description{
Extract attribute classes from the data model.
}
\examples{
\dontrun{
dm_colClasses(data.model = mydatamodel)
}
}
