% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_colClasses.R
\name{dm_colClasses}
\alias{dm_colClasses}
\title{Extract colClasses from data model}
\usage{
dm_colClasses(data.model = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model}
}
\value{
a named vector of the attribute types
}
\description{
Extract colClasses from data model
}
\examples{
dm_colClasses(data.model = mydatamodel)
}
