% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_default.R
\name{dm_default}
\alias{dm_default}
\title{Default Value(s)}
\usage{
dm_default(data.model)
}
\arguments{
\item{data.model}{the data.frame of the data model.}
}
\value{
A data.frame with a 'default' column.
}
\description{
Compute default value(s).
}
\examples{
dm <- data_model(colClasses = c(foo = "numeric"), default = 12)
dm_default(dm)
}
