% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_display.R
\name{dm_display}
\alias{dm_display}
\title{Data Model Display}
\usage{
dm_display(data.model, set = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model.}

\item{set}{an optional character vector with the name of the attributes to set as hidden.}
}
\value{
Either the list of attributes that are hidden or an updated data model if \code{set} is not \code{NULL}
}
\description{
A getter/setter function to manage data model display.
}
\examples{
\dontrun{
dm_display(data.model = mydatamodel, set = c("id", "internal"))
}
}
