% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_filter.R
\name{dm_filter}
\alias{dm_filter}
\title{Extract filters from data model}
\usage{
dm_filter(data.model)
}
\arguments{
\item{data.model}{a data.frame containing the data model}
}
\value{
a named vector of the attribute filters
}
\description{
Extract filters from data model
}
\examples{
\dontrun{
dm_filter(data.model = mydatamodel)
}
}
