% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_integrity.R
\name{dm_integrity}
\alias{dm_integrity}
\title{Data Model Integrity}
\usage{
dm_integrity(data.model, items, fix = FALSE, template = FALSE, n = 100)
}
\arguments{
\item{data.model}{a data.frame of the data model.}

\item{items}{a data.frame of the items.}

\item{fix}{a logical (default = \code{FALSE}) if \code{data.model} should be fixed.}

\item{template}{a logical (default = \code{FALSE}) if attribute template should be used (see details).}

\item{n}{the maximum number (default = 100) of \code{items} rows used to compute an attribute class (see details)}
}
\value{
if \code{data.model} matches with the \code{items}, \code{TRUE} will be returned. Otherwise an updated data model will be returned.
}
\description{
Check the data model integrity.
}
\details{
When \code{fix} is FALSE, an error will be raised if a problem is detected (on first occurrence).

Checks if \code{data.model} matches with the expected data.frame structure.

In case an attribute of \code{items} is missing from \code{data.model}, it will be added to the data model (\code{fix = TRUE}).
When \code{template = TRUE} it will check if the missing attribute(s) matches with en entry in the attribute template
so that they will be added with parameters from the template.
If not found (or when \code{template = FALSE}), it will guess the class of the missing attribute(s) from the \code{n} first rows
of the \code{items}.

In case an extra attribute is found in the \code{data.model} as compared to the \code{items}, it will be dropped from the
data model.
}
\examples{
\dontrun{
feedback <- dm_integrity(mydatamodel, myitems, fix = TRUE)
}
}
