% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_integrity.R
\name{dm_integrity}
\alias{dm_integrity}
\title{Check data model integrity}
\usage{
dm_integrity(data.model, items, template = NULL)
}
\arguments{
\item{data.model}{a \emph{mandatory} data model}

\item{items}{a \emph{mandatory} item data.frame}

\item{template}{an optional data.frame(name = c(...), type = c(...)) to be used to force attribute classes}
}
\value{
if data model matches with the data, TRUE will be returned. Otherwise an updated data model will be returned.
}
\description{
Check data model integrity
}
\details{
In case an attribute is missing from the data model, it will be added with either a type matching the template
one or a guessed one (using the class() function).

In case an extra attribute is found in the data model as compared to the items, it will be dropped from the
data model.
}
\examples{
\dontrun{
feedback <- dm_integrity(mydatamodel, myitems)
if(!is.logical(feedback))
  mydatamodel <- feedback
}
}
