% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_migrate.R
\name{dm_migrate}
\alias{dm_migrate}
\title{Data Model Migration}
\usage{
dm_migrate(data.model)
}
\arguments{
\item{data.model}{a data.frame of the data model.}
}
\value{
The migrated data model or \code{NA} if no migration is required.
}
\description{
Execute the migration procedure on a data model.
}
\examples{
\dontrun{
dm_migrate(data.model)
}
}
