% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_migrate.R
\name{dm_migrate}
\alias{dm_migrate}
\title{Migrate Data Model}
\usage{
dm_migrate(data.model)
}
\arguments{
\item{data.model}{a data.frame of the data model}
}
\value{
the migrated data.model or NA if no migration is required
}
\description{
Migrate Data Model
}
\examples{
\dontrun{
dm_migrate(data.model)
}
}
