% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_names.R
\name{dm_name}
\alias{dm_name}
\title{Data model name}
\usage{
dm_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
the name of the corresponding reactive value
}
\description{
Data model name
}
\examples{
dm_name(id = "mydata")
}
