% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_yaml.R
\name{dm_to_yaml}
\alias{dm_to_yaml}
\title{Data Model To YAML}
\usage{
dm_to_yaml(dm)
}
\arguments{
\item{dm}{a data.frame of the data model}
}
\value{
a list
}
\description{
Convert data model to YAML configuration file.
}
\details{
The input data model should have its version = "0.8.0" otherwise an error
will be raised. Use \code{dm_migrate()} first it needed.
}
\examples{
\dontrun{
dm_to_yaml(dm, name = "my_item")
}
}
