% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_update_attribute.R
\name{dm_update_attribute}
\alias{dm_update_attribute}
\title{Title}
\usage{
dm_update_attribute(
  data.model,
  name,
  default.val = NULL,
  default.fun = NULL,
  filter = FALSE,
  skip = FALSE
)
}
\arguments{
\item{skip}{}
}
\description{
Title
}
\details{
Updating attribute class is not supported by this function (as it requires data migration).

Use of vector to update several attributes is supported as long as length of the different parameters
is either same as name or 1 (then all rows gets same value).
}
\examples{
Use of vector to update several attributes:
dm_update_attribute(data.model = dm, name = c("name","total") , default.val = c("test", 2), default.fun = NA, filter = FALSE, skip = TRUE)
}
