% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_update_attribute.R
\name{dm_update_attribute}
\alias{dm_update_attribute}
\title{Update data model attribute}
\usage{
dm_update_attribute(
  data.model,
  name,
  default.val = NA,
  default.fun = NA,
  filter = FALSE,
  skip = FALSE
)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{name}{a character string of the attribute name}

\item{default.val}{the new default value (default = NULL)}

\item{default.fun}{a character string, the new default function name (default = NULL)}

\item{filter}{a logical to set the filter value (default = FALSE)}

\item{skip}{a logical to set the skip value (default = FALSE)}
}
\value{
the updated data model
}
\description{
Update data model attribute
}
\details{
Updating attribute class is not supported by this function (as it requires data migration).

Use of vector to update several attributes is supported as long as length of the different parameters
is either same as name or 1 (then all rows gets same value).
}
\examples{
\dontrun{
#Use of vector to update several attributes:
dm_update_attribute(data.model = dm,
                    name = c("name","total"),
                    default.val = c("test", 2),
                    default.fun = NA,
                    filter = FALSE,
                    skip = TRUE)
}
}
