% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_update_attribute.R
\name{dm_update_attribute}
\alias{dm_update_attribute}
\title{Update data model attribute}
\usage{
dm_update_attribute(
  data.model,
  name,
  default.val = NULL,
  default.fun = NULL,
  default.arg = NULL,
  skip = NULL
)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{name}{a character string of the attribute name}

\item{default.val}{a character string, the new default value}

\item{default.fun}{a character string, the new default function name}

\item{default.arg}{an optional named vector of arguments, to pass along with the default function.}

\item{skip}{a logical to set the skip value (default = FALSE)}
}
\value{
the updated data model
}
\description{
Update data model attribute
}
\details{
Updating attribute class is not supported by this function (as it requires data migration).

Use of vector to update several attributes is supported as long as length of the different parameters
is either same as name or 1 (then all rows gets same value).

Note that filter should be updated using dm_filter() function
}
\examples{
\dontrun{
#Use of vector to update several attributes:
dm_update_attribute(data.model = dm,
                    name = c("name","total"),
                    default.val = c("test", 2),
                    default.fun = NULL,
                    default.arg = NULL,
                    skip = TRUE)
}
}
\seealso{
\code{\link[=data_model]{data_model()}}
}
