% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_sidebar.R
\name{dynamic_sidebar}
\alias{dynamic_sidebar}
\title{Generate dynamic menuItem}
\usage{
dynamic_sidebar(names)
}
\arguments{
\item{names}{a list of the data model names}
}
\value{
a sidebarMenu menuItem object with one menuSubItem per data model
}
\description{
Generate dynamic menuItem
}
\examples{
\dontrun{
dynamic_sidebar(names = list("data", "data2"))
}
}
