% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_event.R
\name{filter_event}
\alias{filter_event}
\title{Filter Event}
\usage{
filter_event(layer = c("pre", "main"), ...)
}
\arguments{
\item{layer}{the filter layer ("pre" or "main").}

\item{...}{the expression(s) to pass to the filter.}
}
\value{
A list.
}
\description{
Helper function to create filter events to be passed to kitems module server
with the filter argument.
}
\details{
When no expression is passed to \code{...} then the filter layer will be reset.

See the module server function to know how to pass the event to the module
server.

For more details see this vignette:
\code{vignette("filtering", package = "kitems")}
}
\examples{
\dontrun{
# set pre-filtering layer
filter_event(layer = "pre", expr = name == Banana)

# reset main filter
filter_event(layer = "main")
}
}
\seealso{
\code{\link[=kitems]{kitems()}}
}
