% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_event.R
\name{filter_event}
\alias{filter_event}
\title{Create Filter Event}
\usage{
filter_event(layer = c("pre", "main"), ...)
}
\arguments{
\item{layer}{the filter layer ("pre" or "main")}

\item{expr}{the expression(s) to pass to the filter}
}
\value{
an event (list)
}
\description{
Helper function to create filter events to be passed to kitems module server
with the filter argument.
}
\examples{
\dontrun{
# set pre-filtering layer
filter_event(layer = "pre", expr = name == Banana)

# reset main filter
filter_event(layer = "main")
}
}
