% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_view_widget.R
\name{filtered_view_widget}
\alias{filtered_view_widget}
\title{Filtered View Widget}
\usage{
filtered_view_widget(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
a rendered DT data table
}
\description{
Filtered View Widget
}
\examples{
filtered_view_widget(id = "mydata")
}
