% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\title{Item Form}
\usage{
form(attributes, items = NULL, ns)
}
\arguments{
\item{attributes}{a data.frame of the attributes (see details).}

\item{items}{a data.frame of the items (see details).}

\item{ns}{the namespace function, output of \code{shiny::NS()}.}
}
\value{
A list of HTML tags.
}
\description{
Builds the create or update item form.
}
\details{
\code{attributes} is expected to be the output of the \code{default()} or \code{as_default} function.
The data.frame should have the following columns:
"name", "type", "default", "values".

\code{items} are required only when the attribute values defined in the data.model
should be evaluated with data masking support.
}
\examples{
\dontrun{
form(attributes = default(data_model()), ns) << check this!
}
}
