% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasDate.R
\name{hasDate}
\alias{hasDate}
\title{Test date attribute}
\usage{
hasDate(data.model)
}
\arguments{
\item{data.model}{a data.frame containing the data model}
}
\value{
a logical, TRUE if data model has an attribute 'date' (exact match)
}
\description{
Test date attribute
}
\examples{
hasDate(data.model = mydatamodel)
}
