% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R, R/hide_2.R
\name{hide}
\alias{hide}
\alias{display}
\alias{hidden}
\alias{displayed}
\title{Hide Attribute(s)}
\usage{
hide(config, item = get_context(), ...)

hide(config, item = get_context(), ...)

display(config, item = get_context(), ...)

hidden(config, item = get_context())

displayed(config, item = get_context())
}
\arguments{
\item{config}{the config list.}

\item{item}{optional (see details), the name (id) of the item group.}

\item{...}{the name of the attribute(s) to manipulate.}

\item{data.model}{a data.frame of the data model}
}
\value{
a data.frame of the updated data model or a character vector

a config list (setters) or a character vector (getters)
}
\description{
Hide attribute(s) from the data model or get the hidden ones.

A set of verbs to manipulate the display of an attribute.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the hidden attribute's names.

\code{hide()} and \code{display()} are setter functions to tune attribute's behavior.
\code{hidden()} and \code{displayed()} are getter functions to quickly access the attributes
to hide or display in the item table.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
hide(data_model(c(foo = "character")), foo)
hide(data_model(c(foo = "character")), "foo")
hide(data_model(c(foo = "character", bar = "logical")), "foo", bar)

# getter
hide(data_model(c(foo = "character")))
# hide attribute
config <- design(project = "test",
item = "foo") |>
extend(item = "foo", attribute = c(name = "date", type = "Date")) |>
hide(item = "foo", "date")

# allow skipping argument in following calls
item <- "foo"

# hidden attributes
config |> hidden()

# display attribute
config |> display("date")

# displayed attributes
config |> displayed()
}
\seealso{
\code{\link[=display]{display()}} \code{\link[=attribute_update]{attribute_update()}}

\code{\link[=parent.frame]{parent.frame()}}
}
