% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{hide}
\alias{hide}
\title{Hide Attribute(s)}
\usage{
hide(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to hide (quoted or unquoted)}
}
\value{
a data.frame of the updated data model or a character vector
}
\description{
Hide attribute(s) from the data model or get the hidden ones.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the hidden attribute's names.
}
\examples{
hide(data_model(c(foo = "character")), foo)
hide(data_model(c(foo = "character")), "foo")
hide(data_model(c(foo = "character", bar = "logical")), "foo", bar)

# getter
hide(data_model(c(foo = "character")))
}
