% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert Item(s)}
\usage{
insert(values, items)
}
\arguments{
\item{values}{a data.frame of item(s) to insert.}

\item{items}{a data.frame of items where to insert.}
}
\value{
a data.frame of the items
}
\description{
Insert Item(s)
}
\examples{
\dontrun{
insert(values, items)}
}
