% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_truthy.R
\name{is_truthy}
\alias{is_truthy}
\title{Truthy Value}
\usage{
is_truthy(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
a logical
}
\description{
Inspired by Shiny::isTruthy, the purpose here is to determine if a value
is considered as valid to be an attribute value
(or if it will need to be replaced by the attribute defaults)
}
\examples{
is_truthy(12)
}
