% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create.R
\name{item_create}
\alias{item_create}
\title{Create a new item}
\usage{
item_create(values, data.model)
}
\arguments{
\item{values}{a list of values, most likely output values coming from UI inputs}

\item{data.model}{a data.frame, defining the data model}
}
\value{
a data.frame of the new item, coerced to match with colClasses
}
\description{
Create a new item
}
\examples{
\dontrun{
item_create(values, data.model = mydatamodel)
}
}
