% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_delete.R
\name{item_delete}
\alias{item_delete}
\title{Delete item}
\usage{
item_delete(items, id)
}
\arguments{
\item{items}{the reference! of the reactive value carrying the items}

\item{id}{the id of the item to delete}
}
\value{
an updated data.frame of the items
}
\description{
Delete item
}
\examples{
\dontrun{
item_delete(items = myitems, id = 123456789)
}
}
