% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_dialog.R
\name{item_dialog}
\alias{item_dialog}
\title{Item Modal Dialog}
\usage{
item_dialog(
  data.model = NULL,
  items = NULL,
  workflow = c("create", "update", "delete"),
  item = NULL,
  shortcut = FALSE,
  ns
)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{items}{a data.frame of the items}

\item{workflow}{a character string to indicate workflow (see details)}

\item{item}{a data.frame of the item to update (when update = TRUE)}

\item{shortcut}{a logical (default FALSE) if shortcuts should be activated}

\item{ns}{the intented namespace function to use in the dialog}
}
\value{
a modal dialog
}
\description{
Item Modal Dialog
}
\details{
Possible values for workflow are c("create", "update", "delete")
"create" is the default
}
\examples{
\dontrun{
item_dialog(data.model, items, update = FALSE, item = NULL, shortcut = FALSE, ns)
}
}
