% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_form.R
\name{item_form}
\alias{item_form}
\title{Item Form}
\usage{
item_form(data.model, ns)
}
\arguments{
\item{data.model}{the data.frame of the data model.}

\item{ns}{the namespace function, output of \code{shiny::NS()}.}
}
\value{
A list of HTML input objects.
}
\description{
Builds the create or update item form.
}
\details{
Possible values for workflow are "create" (the default) or "update".
}
\examples{
\dontrun{
item_form(data.model = mydata$data_model(), ns)
item_form(data.model = mydata$data_model(), ns)
}
}
