% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_form.R
\name{item_form}
\alias{item_form}
\title{Item Form}
\usage{
item_form(attributes, items = NULL, ns)
}
\arguments{
\item{attributes}{a data.frame of the attributes (see details).}

\item{items}{a data.frame of the items (see details).}

\item{ns}{the namespace function, output of \code{shiny::NS()}.}
}
\value{
A list of HTML input objects.
}
\description{
Builds the create or update item form.
}
\details{
\code{attributes} is expected to be the output of the \code{dm_default()} function.
\code{items} are required only when the attribute values defined in the data.model
should be evaluated with data masking support.
}
\examples{
\dontrun{
item_form(attributes = dm_default(data_model()), ns)
}
}
