% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_form.R
\name{item_form}
\alias{item_form}
\title{Build item input tagList}
\usage{
item_form(ns, item = NULL, update = FALSE, data.model)
}
\arguments{
\item{ns}{the namespace function, output of shiny::NS()}

\item{item}{an optional item (used to set default input values if update = TRUE)}

\item{update}{an optional logical (default = FALSE) to trigger update behavior}

\item{data.model}{a data.frame containing the data model}
}
\value{
a tagList() object containing the attribute inputs
}
\description{
Build item input tagList
}
\details{
Data model skip feature will be used to return inputs only for the skip = FALSE attributes
}
\examples{
\dontrun{
item_form(ns, item = NULL, update = FALSE, data.model = mydatamodel)
item_form(ns, item = myitem, update = TRUE, data.model = mydatamodel)
}
}
