% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_integrity.R
\name{item_integrity}
\alias{item_integrity}
\title{Check items integrity}
\usage{
item_integrity(items, data.model)
}
\arguments{
\item{items}{a data.frame of the items}

\item{data.model}{a data.frame of the data model}
}
\value{
a data.frame of the items, with corrected attribute types
}
\description{
Check items integrity
}
\details{
The function checks if the class of the attributes in items matches with
the one in the data model.
If not, it will coerce the values of the corresponding column to the data model class.
}
\examples{
\dontrun{
items <- item_integrity(items, data.model)
}
}
