% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_load.R
\name{item_load}
\alias{item_load}
\title{Load Items}
\usage{
item_load(connector, col.classes)
}
\arguments{
\item{connector}{a list parameters that will passed to iker::load_data() call.}

\item{col.classes}{a named vector containing the expected column types.}
}
\value{
The data.frame of the items.
}
\description{
Load Items
}
\examples{
\dontrun{
# -- File connector:
item_load(col.classes = c(id = "numeric", date = "Date", comment = "character"),
connector = list(type = "file", path = "path/to/my/data", filename = "mydata.csv"))
}
}
