% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_load.R
\name{item_load}
\alias{item_load}
\title{Load Items}
\usage{
item_load(col.classes, file = NULL, path = NULL)
}
\arguments{
\item{col.classes}{a named vector containing the expected column types.}

\item{file}{an optional file name (including .csv extension).}

\item{path}{an optional path to the file.}
}
\value{
The data.frame of the items.
}
\description{
Load Items
}
\details{
File connector: if file is not \code{NULL}, then data are loaded from the given .csv file.
}
\examples{
\dontrun{
# -- File connector:
item_load(col.classes = c(id = "numeric", date = "Date", comment = "character"),
file = "mydata.csv", path = "path/to/my/data")
}
}
