% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_load.R
\name{item_load}
\alias{item_load}
\title{Load data}
\usage{
item_load(data.model, file = NULL, path = NULL)
}
\arguments{
\item{data.model}{a data.frame containing the data model}

\item{file}{an optional file name (including .csv extension)}

\item{path}{an optional path to the file}
}
\value{
the data (items)
}
\description{
Load data
}
\details{
File connector: if file is not NULL, then data are loaded from the given .csv
}
\examples{
\dontrun{
# -- File connector:
item_load(data.model = mydatamodel, file = "mydata.csv", path = "path/to/my/data")
}
}
