% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_migrate.R
\name{item_migrate}
\alias{item_migrate}
\title{Add attribute to existing items}
\usage{
item_migrate(items, name, type, fill = NA)
}
\arguments{
\item{items}{a data.frame of the items}

\item{name}{a character string of the attribute name}

\item{type}{a character string of the attribute type}

\item{fill}{the value (default = NA) to be used to fill the existing rows (see details)}
}
\value{
the updated items data.frame
}
\description{
Add attribute to existing items
}
\details{
fill will be coerced to the class name provided in type
If a vector is given as input for fill, it will be used: items\link{name} <- value
Make sure the vector length is same as the number of rows, otherwise an error will be raised by R
}
\examples{
\dontrun{
item_migrate(items = myitems, name = "comment", type = "character", fill = "none")
}
}
