% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_save.R
\name{item_save}
\alias{item_save}
\title{Save Items}
\usage{
item_save(data, connector)
}
\arguments{
\item{data}{a data.frame containing the data to be saved.}

\item{connector}{a list parameters that will passed to iker::save_data() call.}
}
\description{
Saves the items data.frame.
}
\examples{
\dontrun{
# -- File connector:
item_save(data = mydata, connector = list(type = "file", file = "path/to/my/data/mydata.csv")
}
}
