% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_save.R
\name{item_save}
\alias{item_save}
\title{Save Items}
\usage{
item_save(data, file = NULL)
}
\arguments{
\item{data}{a data.frame containing the data to be saved.}

\item{file}{the url of the file (including path & .csv extension).}
}
\description{
Saves the items data.frame.
}
\details{
File connector: if file is not \code{NULL}, then data is saved to .csv
}
\examples{
\dontrun{
# -- File connector:
item_save(data = mydata, file = "path/to/my/data/mydata.csv")
}
}
