% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_save.R
\name{item_save}
\alias{item_save}
\title{Save data}
\usage{
item_save(data, file = NULL, path = NULL)
}
\arguments{
\item{data}{a data.frame containing the data to be saved}

\item{file}{an optional file name (including .csv extension)}

\item{path}{an optional path to the file}
}
\description{
Save data
}
\details{
File connector: if file is not NULL, then data is saved to .csv
}
\examples{
\dontrun{
# -- File connector:
item_save(data = mydata, file = "mydata.csv", path = "path/to/my/data")
}
}
