% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_sort.R
\name{item_sort}
\alias{item_sort}
\title{Sort items}
\usage{
item_sort(items, data.model)
}
\arguments{
\item{items}{a data.frame of the items}

\item{data.model}{a data.frame, defining the data model}
}
\value{
a data.frame of the items, sorted based on the data model
}
\description{
Sort items
}
\examples{
\dontrun{
item_sort(items, data.model)
}

}
