% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_update.R
\name{item_update}
\alias{item_update}
\title{Update item}
\usage{
item_update(items, item)
}
\arguments{
\item{items}{the data.frame of the items}

\item{item}{the item to be updated}
}
\value{
the new data.frame of the items
}
\description{
Update item
}
\details{
The item$id value will be used to replace the corresponding item in the items data.frame
}
\examples{
\dontrun{
item_update(items = myitems, item = myupdateditem)
}
}
