% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_update.R
\name{item_update}
\alias{item_update}
\title{Update item}
\usage{
item_update(items, item)
}
\arguments{
\item{items}{a data.frame of the items}

\item{item}{a data.frame of the item(s) to be updated}
}
\value{
the updated data.frame of the items
}
\description{
Update item
}
\details{
The item$id value will be used to replace the corresponding item(s) in the items data.frame
Item rows with id(s) not matching with id(s) in items will be dropped

It is strongly advised to wrap calls to this function inside tryCatch
expressions.
}
\examples{
\dontrun{
item_update(items = myitems, item = myupdateditem)
}
}
