% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{items}
\alias{items}
\title{Get Items}
\usage{
items(datamart, config, item)
}
\arguments{
\item{datamart}{a reactiveValues object holding the items}

\item{config}{the config list}

\item{item}{the name (id) of the item group}
}
\value{
a data.frame of the items or NULL
}
\description{
This function is part of the selective loading mechanism to ensure
items are loaded only when required.
}
\examples{
\dontrun{
# init the datamart
datamart <- reactiveValues()

# get items (it will load them)
datamart |> items("foo")

# get them again (already loaded)
datamart |> items("foo")
}
}
