% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{items_name}
\alias{items_name}
\title{Items Name}
\usage{
items_name(id)
}
\arguments{
\item{id}{the id of the module server instance}
}
\value{
The name of the corresponding items.
}
\description{
Helper function to compute the items name.
}
\examples{
items_name(id = "mydata")
}
