% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems-package.R
\docType{package}
\name{kitems-package}
\alias{kitems-package}
\title{kitems: A Framework To Manage Tabular Items}
\description{
A framework to manage tabular items and a set of tools to implement it within R Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thekangaroofactory.github.io/kitems/}
}

}
\author{
\strong{Maintainer}: Philippe Peret \email{philippe.peret@hotmail.com} (\href{https://orcid.org/0009-0003-9666-7490}{ORCID})

Authors:
\itemize{
  \item Philippe Peret \email{philippe.peret@hotmail.com} (\href{https://orcid.org/0009-0003-9666-7490}{ORCID})
}

}
\keyword{internal}
