% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems.R
\name{kitems}
\alias{kitems}
\title{Module server}
\usage{
kitems(
  id,
  path,
  create = TRUE,
  autosave = TRUE,
  admin = FALSE,
  shortcut = FALSE
)
}
\arguments{
\item{id}{the id to be used for the module server instance}

\item{path}{a path where data model and items will be stored}

\item{create}{a logical whether the data file should be created or not if missing (default = TRUE)}

\item{autosave}{a logical whether the item auto save should be activated or not (default = TRUE)}

\item{admin}{a logical indicating if the admin module server should be launched (default = FALSE)}

\item{shortcut}{a logical should attribute shortcuts be computed when building the item form (default = FALSE)}
}
\description{
Module server
}
\details{
If autosave is FALSE, the item_save function should be used to make the data persistent.
To make the data model persistent, use \link[base]{saveRDS} function. The file name should be
consistent with the output of \link[kitems]{dm_name} function used with \code{id} plus .rds extension.

When admin is FALSE, \link[kitems]{admin_ui} will return an 'empty' layout (tabs with no content)
\link[kitems]{dynamic_sidebar} is not affected by this parameter.
It is expected that those function will not be used when admin = FALSE.
}
\examples{
\dontrun{
kitems(id = "mydata", path = "path/to/my/data",
              create = TRUE, autosave = TRUE)
}
}
