% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_module_server.R
\name{kitemsManager_Server}
\alias{kitemsManager_Server}
\title{Module server}
\usage{
kitemsManager_Server(id, path, create = TRUE, autosave = TRUE)
}
\arguments{
\item{id}{the id to be used for the module server instance}

\item{path}{a path where data model and items will be stored}

\item{create}{a logical whether the data file should be created or not if missing (default = TRUE)}

\item{autosave}{a logical whether the item auto save should be activated or not (default = TRUE)}
}
\description{
Module server
}
\details{
If autosave is FALSE, the save trigger should be used to make the data persistent
}
\examples{
\dontrun{
kitemsManager_Server(id = "mydata", path = "path/to/my/data",
                     create = TRUE, autosave = TRUE)
}
}
