% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitems_module_server.R
\name{kitemsManager_Server}
\alias{kitemsManager_Server}
\title{Title}
\usage{
kitemsManager_Server(
  id,
  r,
  file,
  path,
  data.model = NULL,
  create = TRUE,
  autosave = TRUE
)
}
\arguments{
\item{autosave}{}
}
\description{
Title
}
