% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{name}
\alias{name}
\title{Kitems Names}
\usage{
name(
  id = NULL,
  what = c("item", "dm", "config"),
  file = FALSE,
  url = FALSE,
  backup = FALSE
)
}
\arguments{
\item{id}{the name (id) of the item group.}

\item{what}{the targeted object (item, dm or config).}

\item{file}{a logical if file extension should be added.}

\item{url}{a logical if the url should be returned.}

\item{backup}{a logical if a timestamp should be added.}
}
\value{
a character string.
}
\description{
Helper function to homogenize names across the package.
}
\details{
\code{what} accepts "dm" for backward compatibility reasons (backup & restore old
data.model files before migration to YAML config).

When \code{url} is \code{TRUE}, then \code{file} is considered \code{TRUE} as well.
When \code{backup} is \code{TRUE}, then \code{file} is considered \code{TRUE} as well.

Note that from v0.8.0 url for the config is build from the
\code{R_KITEMS_PATH} environment variable and url for the item groups
is contained in the YAML config.
}
\examples{
# item name
name("foo")
}
