% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{Sort Attributes}
\usage{
organize(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to use for sorting (quoted or unquoted).
Use desc() for descending order (see examples).}
}
\value{
a data.frame of the updated data model or of the sorting parameters
}
\description{
Get or set sorting order of the data.model
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a data.frame of the attributes used to sort the items.

Ranking depends on the position of the attribute.
}
\examples{

dm <- data_model(c(foo = "character", bar = "logical", zoo = "numeric"))

# Sort by attribute foo (ascending order)
organize(dm, foo)

# Sort by attribute foo (descending order)
organize(dm, desc(foo))

# Sort by attribute foo, then bar (descending order)
organize(dm, foo, desc(bar))
}
