% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\alias{organized}
\title{Sort Items Grammar}
\usage{
organize(config, item = get_context(), sort)

organized(config, item = get_context())
}
\arguments{
\item{config}{the config list}

\item{item}{optional (see details), the name (id) of the item group}

\item{sort}{a character string to set the order (see details)}
}
\value{
\code{organize} returns a config list and
\code{organized} returns a character string
}
\description{
Set or get sorting order.
}
\details{
The \code{organize} function accepts a character string for \code{sort} because it is
primarily intended to deal with the Admin Console (which returns
a character string from the user input).

Setting \code{sort = NULL} will reset the item sorting!

Ranking depends on the position of the attribute.
Wrap attribute name by "desc()" to set descending order.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
\dontrun{
organize(config, item = "foo", sort = "date, desc(total)")
organize(config, item = "foo", sort = NULL)
}
}
\seealso{
\code{\link[=parent.frame]{parent.frame()}}
}
