% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_values.R
\name{prepare_values}
\alias{prepare_values}
\title{Prepare Values}
\usage{
prepare_values(values, data.model, update = FALSE)
}
\arguments{
\item{values}{a list of values.}

\item{data.model}{the data.frame of the data model.}

\item{update}{a logical if values are used in the update workflow (default FALSE).}
}
\value{
a data.frame of values
}
\description{
Turn trigger input values into tabular data
}
\details{
values is a named list. The names are used to check the corresponding values
vs the data.model (class, default values if the provided ones are not valid).
The elements in the list must have either length one or same length as the id element.

When an element has length one but the id has several values, all items corresponding
to these ids will be updated with same value. To do so, values will be turned into
a data.frame using as.data.frame ; for this reason, it's strongly advised to wrap
the call into tryCatch as this may fail.
}
\examples{
\dontrun{
prepare_values(values, data.model)
}
}
