% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R, R/refresh_2.R
\name{refresh}
\alias{refresh}
\alias{freeze}
\alias{refreshed}
\alias{frozen}
\title{Refresh Skipped Attribute}
\usage{
refresh(config, item = get_context(), ...)

refresh(config, item = get_context(), ...)

freeze(config, item = get_context(), ...)

refreshed(config, item = get_context())

frozen(config, item = get_context())
}
\arguments{
\item{config}{the config list.}

\item{item}{optional (see details), the name (id) of the item group.}

\item{...}{the name of the attribute(s) to manipulate.}

\item{data.model}{a data.frame of the data model}
}
\value{
a data.frame of the updated data model or a character vector

a config list (setters) or a character vector (getters)
}
\description{
Defines skipped attribute(s) that need a refresh during item update.

A set of verbs to manipulate the refreshed attribute(s) of an item.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the attribute name(s).

\code{refresh()} and \code{freeze()} are setter functions that allow to add or remove attributes from the refreshed ones.
\code{refreshed()} and \code{frozen()} are getter functions to quickly access the skipped attributes
that are refreshed or not during an item update.

Trying to refresh an attribute that is not skipped will be ignored without warning.

When \code{item} is set in the parent frame, then the attribute can be skipped
in the function call.
}
\examples{
refresh(avoid(data_model(c(foo = "character", bar = "logical")), bar), bar)

# attributes not skipped will be ignored
refresh(data_model(c(foo = "character", bar = "logical")), bar)

# getter
refresh(data_model(c(foo = "character", bar = "logical")))
# refresh attribute
config <- design(project = "test",
item = "foo") |>
extend(item = "foo", attribute = c(name = "date", type = "Date")) |>
skip(item = "foo", "date") |>
refresh(item = "foo", "date")

# refreshed attributes
config |> refreshed(item = "foo")

# freeze attribute
config |> freeze(item = "foo", "date")

# frozen attributes
config |> frozen(item = "foo")
}
\seealso{
\code{\link[=avoid]{avoid()}} \code{\link[=attribute_update]{attribute_update()}}

\code{\link[=parent.frame]{parent.frame()}}
}
