% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh Skipped Attribute}
\usage{
refresh(data.model, ...)
}
\arguments{
\item{data.model}{a data.frame of the data model}

\item{...}{the name(s) of the attributes to refresh (quoted or unquoted)}
}
\value{
a data.frame of the updated data model or a character vector
}
\description{
Defines skipped attribute(s) that need a refresh during item update.
}
\details{
When \code{...} is left empty, the function will act as a getter and return
a vector of the attribute name(s).
}
\examples{
refresh(avoid(data_model(c(foo = "character", bar = "logical")), bar), bar)

# attributes not skipped will be ignored
refresh(data_model(c(foo = "character", bar = "logical")), bar)

# getter
refresh(data_model(c(foo = "character", bar = "logical")))
}
\seealso{
\code{\link[=avoid]{avoid()}} \code{\link[=attribute_update]{attribute_update()}}
}
