% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Restore Files}
\usage{
restore(
  type = c("config", "items", "dm"),
  id = NULL,
  path = Sys.getenv("R_KITEMS_PATH")
)
}
\arguments{
\item{type}{the type of file to restore (see details).}

\item{id}{the id for data.model & items.}

\item{path}{optional, the path to the data.}
}
\value{
a logical (see \code{\link[=file.copy]{file.copy()}})
}
\description{
Restore config, data model or items files.
}
\details{
The recommended way to define the \code{path} argument is to set the R_KITEMS_PATH
environment variable.

In case several backup files exist, the newest one will be restored.

Values for \code{type} are "config" (the default), "items" or "dm".
When "config" is used, \code{id} is ignored.
"dm" is kept for compatibility purpose.
}
\examples{
\dontrun{
restore()
restore(type = "items", id = "foo")
}
}
