% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Restore Data Model & Items Files}
\usage{
restore(id, path, type = "items")
}
\arguments{
\item{id}{the kitems id used to create the data model}

\item{path}{the path to the data model}

\item{type}{the type of file to backup. \code{items} (default) or \code{dm}}
}
\description{
Restore Data Model & Items Files
}
\examples{
\dontrun{
restore(id = "mydata", path = "path/to/mydata", type = "items")
restore(id = "mydata", path = "path/to/mydata", type = "dm")
}
}
