% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_insert.R
\name{rows_insert}
\alias{rows_insert}
\title{Insert Item(s)}
\usage{
rows_insert(items, values)
}
\arguments{
\item{items}{a data.frame of the items.}

\item{values}{a list of named values.}
}
\value{
a data.frame of the items
}
\description{
Insert Item(s)
}
\details{
values is a named list. The names are used to check the corresponding values
vs the data.model (class, default values if the provided one are not valid).
The elements in the list must have either length one or same length as the id element.

When an element has length one but the id has several values, all items corresponding
to these ids will be updated with same value. To do so, values will be turned into
a data.frame using as.data.frame ; for this reason, it's strongly advised to wrap
the call into tryCatch as this may fail.
}
\examples{
\dontrun{
rows_insert(items, values)}
}
