% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_update.R
\name{rows_update}
\alias{rows_update}
\title{Update Item(s)}
\usage{
rows_update(items, values, data.model)
}
\arguments{
\item{items}{a data.frame of the items.}

\item{values}{a list of named values.}

\item{data.model}{a data.frame of the data.model.}
}
\value{
a data.frame of the items
}
\description{
Update Item(s)
}
\details{
values is a named list. The names are used to check the corresponding values
vs the data.model (class, default values if the provided one is not valid).
The elements in the list must have either length one or same length as the id element.

When an element has length one but the id has several values, all items corresponding
to these ids will be updated with same value. To do so, values will be turned into
a data.frame using as.data.frame ; for this reason, it's strongly advised to wrap
the call into tryCatch as this may fail.
}
\examples{
\dontrun{
rows_update(items, values, data.model)}
}
