% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Shrink Config}
\usage{
shrink(config, ...)
}
\arguments{
\item{config}{the config list.}

\item{...}{one or several item or attribute instruction.}
}
\value{
a config list.
}
\description{
Shrink Config
}
\details{
Drop an item or an attribute from the config.

The function understands the following instructions:
\itemize{
\item item = foo
\item attribute = c(item = "foo", name = "total")
}

It supports multiple instructions.

Note that it is forbidden to delete the 'id' attribute of an item.
}
\examples{
# build baseline
config <- design(project = "test",
item = "foo") |>
  extend(item = "foo",
         attribute = c(name = "total", type = "integer"))

# drop item
config |>
  shrink(item = "foo")

# drop attribute
config |>
  shrink(attribute = c(item = "foo", name = "total"))

# multiple instructions
config |>
  shrink(attribute = c(item = "foo", name = "total"),
         item = "foo")


}
