% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_event.R
\name{trigger_event}
\alias{trigger_event}
\title{Trigger Event}
\usage{
trigger_event(
  workflow = c("create", "update", "delete"),
  type = c("dialog", "task"),
  values = NULL
)
}
\arguments{
\item{workflow}{the name of the workflow (create, update or delete).}

\item{type}{the name of the action to be performed (dialog or task).}

\item{values}{optional values to create, update or delete an item.}
}
\value{
An event object (list).
}
\description{
Helper function to create events to be passed to kitems module server
with the \code{trigger} argument.
}
\examples{
\dontrun{
# fire create dialog event
trigger_event()
}
}
