% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_widgets.R
\name{update_widget}
\alias{update_widget}
\title{Update Item Button}
\usage{
update_widget(id)
}
\arguments{
\item{id}{the server module id.}
}
\value{
An HTML element that can be included in the UI.
}
\description{
Button to fire the update item dialog.
}
\examples{
\dontrun{
update_widget(id = "mydata")
}
}
