% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate Value(s)}
\usage{
validate(values, data.model, update = FALSE)
}
\arguments{
\item{values}{a list of named values.}

\item{data.model}{the data.frame of the data model (see details).}

\item{update}{whether the id attribute should be checked or not (default FALSE).}
}
\value{
A data.frame of item(s) checked against the data model.
}
\description{
Validate values and turn them into item(s)
}
\details{
This function does not accept the data.model element of the config list as an input.
It should be turned into a data.frame first, using \code{\link[=yaml_to_dm]{yaml_to_dm()}}.

\code{data.model} should contain the following columns:
"name", "type", "default", "class.arg".
}
\examples{
\dontrun{
validate(values, data.model)}
}
