% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_dm.R
\name{yaml_to_dm}
\alias{yaml_to_dm}
\title{Config List To Table}
\usage{
yaml_to_dm(config, ..., item = get_context())
}
\arguments{
\item{config}{the config list (see details).}

\item{...}{the columns to include in the output.}

\item{item}{optional (see details), the name (id) of the item group.}
}
\value{
a data.frame
}
\description{
Turn a config list into a data.frame
}
\details{
The function is used as a bridge between the config list structure and
the expected data.frame input argument for some functions.

It also helps to keep / select the desired columns.
Pass the column names (quoted or unquoted) to \code{...}.
Note that if the function is used inside a package (kitems does), then
the column names should be quoted or R-CMD Check will throw a note about
'no visible binding for global variable' since the column names are not
declared as symbol in the environment.

By default, \code{config} is the global project config list.
\code{item} is required in this case unless it is already set in the parent
frame.

In case the item level config list is provided, it should be up to
the data.model's level. Make sure to explicitly set \code{ìtem = NULL} if
it is set in the parent frame.
}
\examples{
\dontrun{
# get data.model with name and default
yaml_to_dm(config, name, default, item = "foo")
}
}
